+++ control-plane/internal/storage/local.go
@@ -7,12 +7,8 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"net/url"
-	"os"
-	"path/filepath"
 	"reflect"
 	"regexp"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
@@ -22,7 +18,6 @@ import (
 	"github.com/Agent-Field/agentfield/control-plane/pkg/types"
 
 	"github.com/boltdb/bolt"
-	"github.com/jackc/pgx/v5/pgconn"
 	_ "github.com/jackc/pgx/v5/stdlib" // Import pgx driver for PostgreSQL
 	_ "github.com/mattn/go-sqlite3"    // Import sqlite3 driver
 	"gorm.io/gorm"
@@ -493,380 +488,6 @@ type LocalStorage struct {
 }
 
 // NewLocalStorage creates a new instance of LocalStorage.
-func NewLocalStorage(config LocalStorageConfig) *LocalStorage {
-	return &LocalStorage{
-		mode:                      "local",
-		config:                    config,
-		vectorMetric:              VectorDistanceCosine,
-		cache:                     &sync.Map{},
-		subscribers:               make(map[string][]chan types.MemoryChangeEvent),
-		eventBus:                  events.NewExecutionEventBus(),
-		workflowExecutionEventBus: events.NewEventBus[*types.WorkflowExecutionEvent](),
-		executionLogEventBus:      events.NewEventBus[*types.ExecutionLogEntry](),
-	}
-}
-
-// NewPostgresStorage creates a new instance configured for PostgreSQL.
-func NewPostgresStorage(config PostgresStorageConfig) *LocalStorage {
-	return &LocalStorage{
-		mode:                      "postgres",
-		postgresConfig:            config,
-		vectorMetric:              VectorDistanceCosine,
-		cache:                     &sync.Map{},
-		subscribers:               make(map[string][]chan types.MemoryChangeEvent),
-		eventBus:                  events.NewExecutionEventBus(),
-		workflowExecutionEventBus: events.NewEventBus[*types.WorkflowExecutionEvent](),
-		executionLogEventBus:      events.NewEventBus[*types.ExecutionLogEntry](),
-	}
-}
-
-// Initialize sets up the SQLite and BoltDB databases.
-func (ls *LocalStorage) Initialize(ctx context.Context, config StorageConfig) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during initialization: %w", err)
-	}
-
-	mode := config.Mode
-	if mode == "" {
-		mode = "local"
-	}
-
-	ls.mode = mode
-	ls.config = config.Local
-	ls.postgresConfig = config.Postgres
-	ls.vectorConfig = config.Vector.normalized()
-	ls.vectorMetric = parseDistanceMetric(ls.vectorConfig.Distance)
-
-	switch mode {
-	case "local":
-		return ls.initializeSQLite(ctx)
-	case "postgres":
-		return ls.initializePostgres(ctx)
-	default:
-		return fmt.Errorf("unsupported storage mode: %s", mode)
-	}
-}
-
-func (ls *LocalStorage) initializeSQLite(ctx context.Context) error {
-	// Validate that the database path is absolute to prevent files being created in random directories
-	if ls.config.DatabasePath == "" {
-		return fmt.Errorf("database path is empty - please set a valid database path in configuration")
-	}
-
-	// Convert to absolute path if it's relative
-	dbPath := ls.config.DatabasePath
-	if !filepath.IsAbs(dbPath) {
-		absPath, err := filepath.Abs(dbPath)
-		if err != nil {
-			return fmt.Errorf("failed to convert database path to absolute path: %w", err)
-		}
-		logger.Logger.Warn().Msgf("Database path was relative (%s), converted to absolute path: %s", dbPath, absPath)
-		dbPath = absPath
-	}
-
-	// Ensure the directory exists
-	dbDir := filepath.Dir(dbPath)
-	if err := os.MkdirAll(dbDir, 0755); err != nil {
-		return fmt.Errorf("failed to create database directory %s: %w", dbDir, err)
-	}
-
-	logger.Logger.Info().Msgf("Initializing SQLite database at: %s", dbPath)
-
-	busyTimeout := resolveEnvInt("AGENTFIELD_SQLITE_BUSY_TIMEOUT_MS", 60000)
-	if busyTimeout <= 0 {
-		busyTimeout = 60000
-	}
-
-	dsn := fmt.Sprintf("%s?_journal_mode=WAL&_synchronous=NORMAL&_cache_size=10000&_foreign_keys=ON&_busy_timeout=%d&_wal_autocheckpoint=1000&_temp_store=MEMORY&_mmap_size=268435456",
-		dbPath, busyTimeout)
-
-	db, err := sql.Open("sqlite3", dsn)
-	if err != nil {
-		return fmt.Errorf("failed to open SQLite database: %w", err)
-	}
-
-	ls.db = newSQLDatabase(db, "local")
-
-	maxOpen := resolveEnvInt("AGENTFIELD_SQLITE_MAX_OPEN_CONNS", 1)
-	if maxOpen <= 0 {
-		maxOpen = 1
-	}
-	ls.db.SetMaxOpenConns(maxOpen)
-	idleConns := resolveEnvInt("AGENTFIELD_SQLITE_MAX_IDLE_CONNS", 1)
-	if idleConns < 0 {
-		idleConns = 0
-	}
-	ls.db.SetMaxIdleConns(idleConns)
-	ls.db.SetConnMaxLifetime(15 * time.Minute)
-	ls.db.SetConnMaxIdleTime(5 * time.Minute)
-
-	pragmas := []string{
-		"PRAGMA wal_autocheckpoint=1000",
-		"PRAGMA journal_size_limit=67108864",
-		"PRAGMA optimize",
-	}
-
-	for _, pragma := range pragmas {
-		if _, err := ls.db.Exec(pragma); err != nil {
-			return fmt.Errorf("failed to set pragma %s: %w", pragma, err)
-		}
-	}
-
-	if err := ls.initGormDB(); err != nil {
-		return fmt.Errorf("failed to initialize gorm: %w", err)
-	}
-
-	go ls.periodicWALCheckpoint()
-
-	kvStore, err := bolt.Open(ls.config.KVStorePath, 0600, &bolt.Options{Timeout: 1 * time.Second})
-	if err != nil {
-		return fmt.Errorf("failed to open BoltDB database: %w", err)
-	}
-	ls.kvStore = kvStore
-
-	if err := ls.createSchema(ctx); err != nil {
-		return fmt.Errorf("failed to create local storage schema: %w", err)
-	}
-
-	return nil
-}
-
-func resolveEnvInt(key string, fallback int) int {
-	raw := strings.TrimSpace(os.Getenv(key))
-	if raw == "" {
-		return fallback
-	}
-	value, err := strconv.Atoi(raw)
-	if err != nil {
-		logger.Logger.Warn().Msgf("Invalid integer for %s=%s, using fallback %d", key, raw, fallback)
-		return fallback
-	}
-	return value
-}
-
-func (ls *LocalStorage) initializePostgres(ctx context.Context) error {
-	cfg := ls.postgresConfig
-	dsn := strings.TrimSpace(cfg.DSN)
-	if dsn == "" {
-		dsn = strings.TrimSpace(cfg.URL)
-	}
-
-	if dsn == "" {
-		if cfg.Host == "" {
-			return fmt.Errorf("postgres configuration requires either a connection string or host information")
-		}
-
-		if cfg.Port == 0 {
-			cfg.Port = 5432
-		}
-
-		if cfg.User == "" {
-			return fmt.Errorf("postgres configuration requires a user when host is specified")
-		}
-
-		if cfg.Database == "" {
-			return fmt.Errorf("postgres configuration requires a database when host is specified")
-		}
-
-		if cfg.SSLMode == "" {
-			cfg.SSLMode = "disable"
-		}
-
-		pgURL := &url.URL{
-			Scheme: "postgres",
-			Host:   fmt.Sprintf("%s:%d", cfg.Host, cfg.Port),
-			Path:   "/" + strings.TrimPrefix(cfg.Database, "/"),
-		}
-
-		if cfg.Password != "" {
-			pgURL.User = url.UserPassword(cfg.User, cfg.Password)
-		} else {
-			pgURL.User = url.User(cfg.User)
-		}
-
-		query := pgURL.Query()
-		if cfg.SSLMode != "" {
-			query.Set("sslmode", cfg.SSLMode)
-		}
-		if len(query) > 0 {
-			pgURL.RawQuery = query.Encode()
-		}
-
-		dsn = pgURL.String()
-	}
-
-	if cfg.SSLMode == "" {
-		cfg.SSLMode = "disable"
-	}
-
-	cfg.DSN = dsn
-	cfg.URL = dsn
-	ls.postgresConfig = cfg
-
-	db, err := sql.Open("pgx", dsn)
-	if err != nil {
-		return fmt.Errorf("failed to open PostgreSQL database: %w", err)
-	}
-
-	sqlDB := newSQLDatabase(db, "postgres")
-	ls.applyPostgresConnectionSettings(sqlDB, cfg)
-
-	if err := sqlDB.PingContext(ctx); err != nil {
-		if isPostgresDatabaseMissingError(err) {
-			_ = sqlDB.Close()
-			if err := ls.ensurePostgresDatabaseExists(ctx, cfg); err != nil {
-				return fmt.Errorf("failed to create postgres database: %w", err)
-			}
-
-			db, err = sql.Open("pgx", cfg.DSN)
-			if err != nil {
-				return fmt.Errorf("failed to open PostgreSQL database after creation: %w", err)
-			}
-
-			sqlDB = newSQLDatabase(db, "postgres")
-			ls.applyPostgresConnectionSettings(sqlDB, cfg)
-
-			if err := sqlDB.PingContext(ctx); err != nil {
-				_ = sqlDB.Close()
-				return fmt.Errorf("failed to ping PostgreSQL database after creation: %w", err)
-			}
-		} else {
-			_ = sqlDB.Close()
-			return fmt.Errorf("failed to ping PostgreSQL database: %w", err)
-		}
-	}
-
-	ls.db = sqlDB
-
-	if err := ls.initGormDB(); err != nil {
-		return fmt.Errorf("failed to initialize gorm for postgres: %w", err)
-	}
-
-	if err := ls.createSchema(ctx); err != nil {
-		return fmt.Errorf("failed to create postgres storage schema: %w", err)
-	}
-
-	return nil
-}
-
-func (ls *LocalStorage) applyPostgresConnectionSettings(db *sqlDatabase, cfg PostgresStorageConfig) {
-	if db == nil {
-		return
-	}
-
-	maxOpen := cfg.MaxOpenConns
-	if maxOpen <= 0 {
-		maxOpen = 25
-	}
-	maxIdle := cfg.MaxIdleConns
-	if maxIdle <= 0 {
-		maxIdle = 5
-	}
-
-	db.SetMaxOpenConns(maxOpen)
-	db.SetMaxIdleConns(maxIdle)
-
-	if cfg.ConnMaxLifetime > 0 {
-		db.SetConnMaxLifetime(cfg.ConnMaxLifetime)
-	} else {
-		db.SetConnMaxLifetime(30 * time.Minute)
-	}
-}
-
-func isPostgresDatabaseMissingError(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	var pgErr *pgconn.PgError
-	if errors.As(err, &pgErr) {
-		return pgErr.Code == "3D000" // undefined_database
-	}
-
-	return strings.Contains(err.Error(), "does not exist")
-}
-
-func isPostgresDatabaseAlreadyExistsError(err error) bool {
-	if err == nil {
-		return false
-	}
-
-	var pgErr *pgconn.PgError
-	if errors.As(err, &pgErr) {
-		return pgErr.Code == "42P04" // duplicate_database
-	}
-
-	return strings.Contains(err.Error(), "already exists")
-}
-
-func (ls *LocalStorage) ensurePostgresDatabaseExists(ctx context.Context, cfg PostgresStorageConfig) error {
-	dsn := strings.TrimSpace(cfg.DSN)
-	if dsn == "" {
-		return fmt.Errorf("postgres DSN is required to create database")
-	}
-
-	parsed, err := url.Parse(dsn)
-	if err != nil {
-		return fmt.Errorf("failed to parse postgres DSN: %w", err)
-	}
-
-	dbName := strings.TrimPrefix(parsed.Path, "/")
-	dbName = strings.TrimSpace(dbName)
-	if dbName == "" {
-		return fmt.Errorf("postgres DSN must specify a database name")
-	}
-
-	adminDatabase := strings.TrimSpace(cfg.AdminDatabase)
-	if adminDatabase == "" {
-		adminDatabase = "postgres"
-	}
-
-	parsed.Path = "/" + adminDatabase
-	adminDSN := parsed.String()
-
-	adminDB, err := sql.Open("pgx", adminDSN)
-	if err != nil {
-		return fmt.Errorf("failed to open postgres admin connection: %w", err)
-	}
-	defer adminDB.Close()
-
-	if err := adminDB.PingContext(ctx); err != nil {
-		return fmt.Errorf("failed to ping postgres admin database: %w", err)
-	}
-
-	createStmt := fmt.Sprintf("CREATE DATABASE %s", quotePostgresIdentifier(dbName))
-	if _, err := adminDB.ExecContext(ctx, createStmt); err != nil {
-		if isPostgresDatabaseAlreadyExistsError(err) {
-			return nil
-		}
-		return fmt.Errorf("failed to create postgres database '%s': %w", dbName, err)
-	}
-
-	return nil
-}
-
-func quotePostgresIdentifier(identifier string) string {
-	escaped := strings.ReplaceAll(identifier, "\"", "\"\"")
-	return "\"" + escaped + "\""
-}
-
-// periodicWALCheckpoint runs periodic WAL checkpoints to prevent WAL file buildup
-// SQLite WAL mode handles write coordination internally, so no additional locking needed
-func (ls *LocalStorage) periodicWALCheckpoint() {
-	ticker := time.NewTicker(30 * time.Second)
-	defer ticker.Stop()
-
-	for range ticker.C {
-		if ls.db != nil {
-			// Perform checkpoint - SQLite WAL mode handles coordination internally
-			_, err := ls.db.Exec("PRAGMA wal_checkpoint(PASSIVE)")
-			if err != nil {
-				// Log error but don't stop the checkpoint process
-				fmt.Printf("WAL checkpoint failed: %v\n", err)
-			}
-		}
-	}
-}
 
 // createSchema ensures the SQLite schema, indexes, and supporting buckets exist.
 func (ls *LocalStorage) createSchema(ctx context.Context) error {
@@ -6293,2717 +5914,4 @@ func (ls *LocalStorage) GetExecutionLogEventBus() *events.EventBus[*types.Execut
 }
 
 // AgentField Server DID operations
-func (ls *LocalStorage) StoreAgentFieldServerDID(ctx context.Context, agentfieldServerID, rootDID string, masterSeed []byte, createdAt, lastKeyRotation time.Time) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store af server DID: %w", err)
-	}
 
-	// Validate input parameters
-	if agentfieldServerID == "" {
-		return &ValidationError{
-			Field:   "agentfield_server_id",
-			Value:   agentfieldServerID,
-			Reason:  "af server ID cannot be empty",
-			Context: "StoreAgentFieldServerDID",
-		}
-	}
-	if rootDID == "" {
-		return &ValidationError{
-			Field:   "root_did",
-			Value:   rootDID,
-			Reason:  "root DID cannot be empty",
-			Context: "StoreAgentFieldServerDID",
-		}
-	}
-	if len(masterSeed) == 0 {
-		return &ValidationError{
-			Field:   "master_seed",
-			Value:   "<encrypted>",
-			Reason:  "master seed cannot be empty",
-			Context: "StoreAgentFieldServerDID",
-		}
-	}
-
-	// Use transaction for data consistency
-	tx, err := ls.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			rollbackTx(tx, "StoreAgentFieldServerDID")
-		}
-	}()
-
-	// Execute with retry logic
-	err = ls.retryOnConstraintFailure(ctx, func() error {
-		query := `
-                        INSERT OR REPLACE INTO did_registry (agentfield_server_id, root_did, master_seed_encrypted, created_at, last_key_rotation, total_dids)
-                        VALUES (?, ?, ?, ?, ?, 0)
-                `
-		if ls.mode == "postgres" {
-			query = `
-                                INSERT INTO did_registry (agentfield_server_id, root_did, master_seed_encrypted, created_at, last_key_rotation, total_dids)
-                                VALUES (?, ?, ?, ?, ?, 0)
-                                ON CONFLICT (agentfield_server_id) DO UPDATE SET
-                                        root_did = EXCLUDED.root_did,
-                                        master_seed_encrypted = EXCLUDED.master_seed_encrypted,
-                                        created_at = EXCLUDED.created_at,
-                                        last_key_rotation = EXCLUDED.last_key_rotation,
-                                        total_dids = did_registry.total_dids
-                        `
-		}
-		_, execErr := tx.ExecContext(ctx, query, agentfieldServerID, rootDID, masterSeed, createdAt, lastKeyRotation)
-		if execErr != nil {
-			return fmt.Errorf("failed to store af server DID: %w", execErr)
-		}
-		return nil
-	}, 3) // Retry up to 3 times for transient errors
-
-	if err != nil {
-		return err
-	}
-
-	// Commit transaction
-	if err = tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	logger.Logger.Debug().Msgf("Successfully stored af server DID: agentfield_server_id=%s, root_did=%s", agentfieldServerID, rootDID)
-	return nil
-}
-
-// StoreAgentDIDWithComponents stores an agent DID along with its component DIDs in a single transaction
-func (ls *LocalStorage) StoreAgentDIDWithComponents(ctx context.Context, agentID, agentDID, agentfieldServerDID, publicKeyJWK string, derivationIndex int, components []ComponentDIDRequest) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store agent DID with components: %w", err)
-	}
-
-	// Pre-storage validation
-	if err := ls.validateAgentFieldServerExists(ctx, agentfieldServerDID); err != nil {
-		return fmt.Errorf("pre-storage validation failed: %w", err)
-	}
-
-	// Use transaction for data consistency across all operations
-	tx, err := ls.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			rollbackTx(tx, "StoreAgentDIDWithComponents")
-		}
-	}()
-
-	// Store agent DID first
-	err = ls.retryOnConstraintFailure(ctx, func() error {
-		query := `
-			INSERT INTO agent_dids (
-				agent_node_id, did, agentfield_server_id, public_key_jwk, derivation_path, registered_at, status
-			) VALUES (?, ?, ?, ?, ?, ?, ?)`
-
-		derivationPath := fmt.Sprintf("m/44'/0'/0'/%d", derivationIndex)
-		_, execErr := tx.ExecContext(ctx, query, agentID, agentDID, agentfieldServerDID, publicKeyJWK, derivationPath, time.Now(), "active")
-		if execErr != nil {
-			if strings.Contains(execErr.Error(), "UNIQUE constraint failed") || strings.Contains(execErr.Error(), "agent_dids") {
-				return &DuplicateDIDError{
-					DID:  fmt.Sprintf("agent:%s@%s", agentID, agentfieldServerDID),
-					Type: "agent",
-				}
-			}
-			if strings.Contains(execErr.Error(), "FOREIGN KEY constraint failed") {
-				return &ForeignKeyConstraintError{
-					Table:           "agent_dids",
-					Column:          "agentfield_server_id",
-					ReferencedTable: "did_registry",
-					ReferencedValue: agentfieldServerDID,
-					Operation:       "INSERT",
-				}
-			}
-			return fmt.Errorf("failed to store agent DID: %w", execErr)
-		}
-		return nil
-	}, 3)
-
-	if err != nil {
-		var dupErr *DuplicateDIDError
-		if errors.As(err, &dupErr) {
-			return dupErr
-		}
-		return fmt.Errorf("failed to store agent DID: %w", err)
-	}
-
-	// Store component DIDs
-	for i, component := range components {
-		err = ls.retryOnConstraintFailure(ctx, func() error {
-			query := `
-				INSERT INTO component_dids (
-					did, agent_did, component_type, function_name, public_key_jwk, derivation_path
-				) VALUES (?, ?, ?, ?, ?, ?)`
-
-			derivationPath := fmt.Sprintf("m/44'/0'/0'/%d", component.DerivationIndex)
-			_, execErr := tx.ExecContext(ctx, query, component.ComponentDID, agentDID, component.ComponentType, component.ComponentName, component.PublicKeyJWK, derivationPath)
-			if execErr != nil {
-				if strings.Contains(execErr.Error(), "UNIQUE constraint failed") || strings.Contains(execErr.Error(), "component_dids") {
-					return &DuplicateDIDError{
-						DID:  fmt.Sprintf("component:%s/%s@%s", component.ComponentType, component.ComponentName, agentDID),
-						Type: "component",
-					}
-				}
-				if strings.Contains(execErr.Error(), "FOREIGN KEY constraint failed") {
-					return &ForeignKeyConstraintError{
-						Table:           "component_dids",
-						Column:          "agent_did",
-						ReferencedTable: "agent_dids",
-						ReferencedValue: agentDID,
-						Operation:       "INSERT",
-					}
-				}
-				return fmt.Errorf("failed to store component DID %d: %w", i, execErr)
-			}
-			return nil
-		}, 3)
-
-		if err != nil {
-			var dupErr *DuplicateDIDError
-			if errors.As(err, &dupErr) {
-				return dupErr
-			}
-			return fmt.Errorf("failed to store component DID %d (%s): %w", i, component.ComponentName, err)
-		}
-	}
-
-	// Commit transaction
-	if err = tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	logger.Logger.Debug().Msgf("Successfully stored agent DID with %d components: agent_id=%s, did=%s", len(components), agentID, agentDID)
-	return nil
-}
-
-func (ls *LocalStorage) GetAgentFieldServerDID(ctx context.Context, agentfieldServerID string) (*types.AgentFieldServerDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get af server DID: %w", err)
-	}
-
-	query := `
-		SELECT agentfield_server_id, root_did, master_seed_encrypted, created_at, last_key_rotation
-		FROM did_registry WHERE agentfield_server_id = ?
-	`
-	row := ls.db.QueryRowContext(ctx, query, agentfieldServerID)
-	info := &types.AgentFieldServerDIDInfo{}
-
-	err := row.Scan(&info.AgentFieldServerID, &info.RootDID, &info.MasterSeed, &info.CreatedAt, &info.LastKeyRotation)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, nil // Return nil, nil for "not found"
-		}
-		return nil, fmt.Errorf("failed to get af server DID: %w", err)
-	}
-	return info, nil
-}
-
-func (ls *LocalStorage) ListAgentFieldServerDIDs(ctx context.Context) ([]*types.AgentFieldServerDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list af server DIDs: %w", err)
-	}
-
-	query := `
-		SELECT agentfield_server_id, root_did, master_seed_encrypted, created_at, last_key_rotation
-		FROM did_registry ORDER BY created_at DESC
-	`
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list af server DIDs: %w", err)
-	}
-	defer rows.Close()
-
-	var infos []*types.AgentFieldServerDIDInfo
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during af server DID list iteration: %w", err)
-		}
-
-		info := &types.AgentFieldServerDIDInfo{}
-		err := rows.Scan(&info.AgentFieldServerID, &info.RootDID, &info.MasterSeed, &info.CreatedAt, &info.LastKeyRotation)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan af server DID: %w", err)
-		}
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-// DID Registry operations
-func (ls *LocalStorage) StoreDID(ctx context.Context, did string, didDocument, publicKey, privateKeyRef, derivationPath string) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store DID: %w", err)
-	}
-
-	// INSERT-only query - no ON CONFLICT clause for security
-	query := `
-		INSERT INTO did_registry (
-			did, did_document, public_key, private_key_ref, derivation_path,
-			created_at, updated_at, status
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?)`
-
-	now := time.Now()
-	_, err := ls.db.ExecContext(ctx, query, did, didDocument, publicKey, privateKeyRef, derivationPath, now, now, "active")
-	if err != nil {
-		// Check if this is a unique constraint violation (duplicate DID)
-		if strings.Contains(err.Error(), "UNIQUE constraint failed") || strings.Contains(err.Error(), "did_registry.did") {
-			logger.Logger.Warn().Msgf("Duplicate DID registry entry detected: %s", did)
-			return &DuplicateDIDError{
-				DID:  did,
-				Type: "registry",
-			}
-		}
-		return fmt.Errorf("failed to store DID: %w", err)
-	}
-
-	logger.Logger.Debug().Msgf("Successfully stored DID registry entry: %s", did)
-	return nil
-}
-
-func (ls *LocalStorage) GetDID(ctx context.Context, did string) (*types.DIDRegistryEntry, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get DID: %w", err)
-	}
-
-	query := `
-		SELECT did, did_document, public_key, private_key_ref, derivation_path,
-			   created_at, updated_at, status
-		FROM did_registry WHERE did = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, did)
-	entry := &types.DIDRegistryEntry{}
-
-	err := row.Scan(&entry.DID, &entry.DIDDocument, &entry.PublicKey, &entry.PrivateKeyRef,
-		&entry.DerivationPath, &entry.CreatedAt, &entry.UpdatedAt, &entry.Status)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("DID %s not found", did)
-		}
-		return nil, fmt.Errorf("failed to get DID: %w", err)
-	}
-	return entry, nil
-}
-
-func (ls *LocalStorage) ListDIDs(ctx context.Context) ([]*types.DIDRegistryEntry, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list DIDs: %w", err)
-	}
-
-	query := `
-		SELECT did, did_document, public_key, private_key_ref, derivation_path,
-			   created_at, updated_at, status
-		FROM did_registry ORDER BY created_at DESC`
-
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list DIDs: %w", err)
-	}
-	defer rows.Close()
-
-	var entries []*types.DIDRegistryEntry
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during DID list iteration: %w", err)
-		}
-
-		entry := &types.DIDRegistryEntry{}
-		err := rows.Scan(&entry.DID, &entry.DIDDocument, &entry.PublicKey, &entry.PrivateKeyRef,
-			&entry.DerivationPath, &entry.CreatedAt, &entry.UpdatedAt, &entry.Status)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan DID entry: %w", err)
-		}
-		entries = append(entries, entry)
-	}
-	return entries, nil
-}
-
-// validateAgentFieldServerExists checks if a af server registry exists
-func (ls *LocalStorage) validateAgentFieldServerExists(ctx context.Context, agentfieldServerID string) error {
-	if agentfieldServerID == "" {
-		return &ValidationError{
-			Field:   "agentfield_server_id",
-			Value:   agentfieldServerID,
-			Reason:  "af server ID cannot be empty",
-			Context: "pre-storage validation",
-		}
-	}
-
-	query := `SELECT 1 FROM did_registry WHERE agentfield_server_id = ? LIMIT 1`
-	var exists int
-	err := ls.db.QueryRowContext(ctx, query, agentfieldServerID).Scan(&exists)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return &ForeignKeyConstraintError{
-				Table:           "agent_dids",
-				Column:          "agentfield_server_id",
-				ReferencedTable: "did_registry",
-				ReferencedValue: agentfieldServerID,
-				Operation:       "INSERT",
-			}
-		}
-		return fmt.Errorf("failed to validate af server existence: %w", err)
-	}
-	return nil
-}
-
-// validateAgentDIDExists checks if an agent DID exists
-func (ls *LocalStorage) validateAgentDIDExists(ctx context.Context, agentDID string) error {
-	if agentDID == "" {
-		return &ValidationError{
-			Field:   "agent_did",
-			Value:   agentDID,
-			Reason:  "agent DID cannot be empty",
-			Context: "pre-storage validation",
-		}
-	}
-
-	query := `SELECT 1 FROM agent_dids WHERE did = ? LIMIT 1`
-	var exists int
-	err := ls.db.QueryRowContext(ctx, query, agentDID).Scan(&exists)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return &ForeignKeyConstraintError{
-				Table:           "component_dids",
-				Column:          "agent_did",
-				ReferencedTable: "agent_dids",
-				ReferencedValue: agentDID,
-				Operation:       "INSERT",
-			}
-		}
-		return fmt.Errorf("failed to validate agent DID existence: %w", err)
-	}
-	return nil
-}
-
-// retryOnConstraintFailure executes a function with retry logic for transient constraint issues
-func (ls *LocalStorage) retryOnConstraintFailure(ctx context.Context, operation func() error, maxRetries int) error {
-	var lastErr error
-	for attempt := 0; attempt <= maxRetries; attempt++ {
-		if err := ctx.Err(); err != nil {
-			return fmt.Errorf("context cancelled during retry attempt %d: %w", attempt, err)
-		}
-
-		lastErr = operation()
-		if lastErr == nil {
-			return nil
-		}
-
-		// Don't retry on validation errors or permanent constraint violations
-		if _, isValidationErr := lastErr.(*ValidationError); isValidationErr {
-			return lastErr
-		}
-		if _, isFKErr := lastErr.(*ForeignKeyConstraintError); isFKErr {
-			return lastErr
-		}
-		if _, isDuplicateErr := lastErr.(*DuplicateDIDError); isDuplicateErr {
-			return lastErr
-		}
-
-		// Only retry on database-level transient errors
-		if strings.Contains(lastErr.Error(), "database is locked") ||
-			strings.Contains(lastErr.Error(), "SQLITE_BUSY") ||
-			strings.Contains(lastErr.Error(), "database is temporarily unavailable") {
-			if attempt < maxRetries {
-				// Exponential backoff: 10ms, 20ms, 40ms
-				backoff := time.Duration(10*(1<<attempt)) * time.Millisecond
-				time.Sleep(backoff)
-				continue
-			}
-		}
-
-		// For other errors, don't retry
-		return lastErr
-	}
-	return lastErr
-}
-
-// Agent DID operations
-func (ls *LocalStorage) StoreAgentDID(ctx context.Context, agentID, agentDID, agentfieldServerDID, publicKeyJWK string, derivationIndex int) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store agent DID: %w", err)
-	}
-
-	// Pre-storage validation
-	if err := ls.validateAgentFieldServerExists(ctx, agentfieldServerDID); err != nil {
-		return fmt.Errorf("pre-storage validation failed: %w", err)
-	}
-
-	// Validate input parameters
-	if agentID == "" {
-		return &ValidationError{
-			Field:   "agent_node_id",
-			Value:   agentID,
-			Reason:  "agent ID cannot be empty",
-			Context: "StoreAgentDID",
-		}
-	}
-	if agentDID == "" {
-		return &ValidationError{
-			Field:   "did",
-			Value:   agentDID,
-			Reason:  "agent DID cannot be empty",
-			Context: "StoreAgentDID",
-		}
-	}
-	if publicKeyJWK == "" {
-		return &ValidationError{
-			Field:   "public_key_jwk",
-			Value:   publicKeyJWK,
-			Reason:  "public key JWK cannot be empty",
-			Context: "StoreAgentDID",
-		}
-	}
-
-	// Use transaction for data consistency
-	tx, err := ls.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			rollbackTx(tx, "StoreAgentDID")
-		}
-	}()
-
-	// Execute with retry logic
-	err = ls.retryOnConstraintFailure(ctx, func() error {
-		// INSERT-only query - no ON CONFLICT clause for security
-		query := `
-			INSERT INTO agent_dids (
-				agent_node_id, did, agentfield_server_id, public_key_jwk, derivation_path, registered_at, status
-			) VALUES (?, ?, ?, ?, ?, ?, ?)`
-
-		derivationPath := fmt.Sprintf("m/44'/0'/0'/%d", derivationIndex)
-		_, execErr := tx.ExecContext(ctx, query, agentID, agentDID, agentfieldServerDID, publicKeyJWK, derivationPath, time.Now(), "active")
-		if execErr != nil {
-			// Check if this is a unique constraint violation (duplicate agent DID)
-			if strings.Contains(execErr.Error(), "UNIQUE constraint failed") || strings.Contains(execErr.Error(), "agent_dids") {
-				logger.Logger.Warn().Msgf("Duplicate agent DID entry detected: agent_id=%s, agentfield_server_id=%s", agentID, agentfieldServerDID)
-				return &DuplicateDIDError{
-					DID:  fmt.Sprintf("agent:%s@%s", agentID, agentfieldServerDID),
-					Type: "agent",
-				}
-			}
-			// Check for foreign key constraint violations
-			if strings.Contains(execErr.Error(), "FOREIGN KEY constraint failed") {
-				return &ForeignKeyConstraintError{
-					Table:           "agent_dids",
-					Column:          "agentfield_server_id",
-					ReferencedTable: "did_registry",
-					ReferencedValue: agentfieldServerDID,
-					Operation:       "INSERT",
-				}
-			}
-			return fmt.Errorf("failed to store agent DID: %w", execErr)
-		}
-		return nil
-	}, 3) // Retry up to 3 times for transient errors
-
-	if err != nil {
-		return err
-	}
-
-	// Commit transaction
-	if err = tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	logger.Logger.Debug().Msgf("Successfully stored agent DID entry: agent_id=%s, did=%s", agentID, agentDID)
-	return nil
-}
-
-func (ls *LocalStorage) GetAgentDID(ctx context.Context, agentID string) (*types.AgentDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get agent DID: %w", err)
-	}
-
-	query := `
-		SELECT agent_node_id, did, agentfield_server_id, public_key_jwk, derivation_path,
-		       reasoners, skills, status, registered_at
-		FROM agent_dids WHERE agent_node_id = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, agentID)
-	info := &types.AgentDIDInfo{}
-
-	var reasonersJSON, skillsJSON, publicKeyJWK string
-	err := row.Scan(&info.AgentNodeID, &info.DID, &info.AgentFieldServerID, &publicKeyJWK,
-		&info.DerivationPath, &reasonersJSON, &skillsJSON, &info.Status, &info.RegisteredAt)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("agent DID for %s not found", agentID)
-		}
-		return nil, fmt.Errorf("failed to get agent DID: %w", err)
-	}
-	info.PublicKeyJWK = json.RawMessage(publicKeyJWK)
-
-	// Parse JSON fields
-	if reasonersJSON != "" {
-		if err := json.Unmarshal([]byte(reasonersJSON), &info.Reasoners); err != nil {
-			return nil, fmt.Errorf("failed to parse reasoners JSON: %w", err)
-		}
-	} else {
-		info.Reasoners = make(map[string]types.ReasonerDIDInfo)
-	}
-
-	if skillsJSON != "" {
-		if err := json.Unmarshal([]byte(skillsJSON), &info.Skills); err != nil {
-			return nil, fmt.Errorf("failed to parse skills JSON: %w", err)
-		}
-	} else {
-		info.Skills = make(map[string]types.SkillDIDInfo)
-	}
-
-	return info, nil
-}
-
-func (ls *LocalStorage) ListAgentDIDs(ctx context.Context) ([]*types.AgentDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list agent DIDs: %w", err)
-	}
-
-	query := `
-		SELECT agent_node_id, did, agentfield_server_id, public_key_jwk, derivation_path,
-		       reasoners, skills, status, registered_at
-		FROM agent_dids ORDER BY registered_at DESC`
-
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list agent DIDs: %w", err)
-	}
-	defer rows.Close()
-
-	var infos []*types.AgentDIDInfo
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during agent DID list iteration: %w", err)
-		}
-
-		info := &types.AgentDIDInfo{}
-		var reasonersJSON, skillsJSON, publicKeyJWK string
-		err := rows.Scan(&info.AgentNodeID, &info.DID, &info.AgentFieldServerID, &publicKeyJWK,
-			&info.DerivationPath, &reasonersJSON, &skillsJSON, &info.Status, &info.RegisteredAt)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan agent DID: %w", err)
-		}
-		info.PublicKeyJWK = json.RawMessage(publicKeyJWK)
-
-		// Parse JSON fields
-		if reasonersJSON != "" {
-			if err := json.Unmarshal([]byte(reasonersJSON), &info.Reasoners); err != nil {
-				return nil, fmt.Errorf("failed to parse reasoners JSON: %w", err)
-			}
-		} else {
-			info.Reasoners = make(map[string]types.ReasonerDIDInfo)
-		}
-
-		if skillsJSON != "" {
-			if err := json.Unmarshal([]byte(skillsJSON), &info.Skills); err != nil {
-				return nil, fmt.Errorf("failed to parse skills JSON: %w", err)
-			}
-		} else {
-			info.Skills = make(map[string]types.SkillDIDInfo)
-		}
-
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-// Component DID operations
-func (ls *LocalStorage) StoreComponentDID(ctx context.Context, componentID, componentDID, agentDID, componentType, componentName string, derivationIndex int) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store component DID: %w", err)
-	}
-
-	// Pre-storage validation
-	if err := ls.validateAgentDIDExists(ctx, agentDID); err != nil {
-		return fmt.Errorf("pre-storage validation failed: %w", err)
-	}
-
-	// Validate input parameters
-	if componentDID == "" {
-		return &ValidationError{
-			Field:   "component_did",
-			Value:   componentDID,
-			Reason:  "component DID cannot be empty",
-			Context: "StoreComponentDID",
-		}
-	}
-	if componentType == "" {
-		return &ValidationError{
-			Field:   "component_type",
-			Value:   componentType,
-			Reason:  "component type cannot be empty",
-			Context: "StoreComponentDID",
-		}
-	}
-	if componentName == "" {
-		return &ValidationError{
-			Field:   "component_name",
-			Value:   componentName,
-			Reason:  "component name cannot be empty",
-			Context: "StoreComponentDID",
-		}
-	}
-	// Validate component type
-	validTypes := map[string]bool{"reasoner": true, "skill": true}
-	if !validTypes[componentType] {
-		return &ValidationError{
-			Field:   "component_type",
-			Value:   componentType,
-			Reason:  "component type must be 'reasoner' or 'skill'",
-			Context: "StoreComponentDID",
-		}
-	}
-
-	// Use transaction for data consistency
-	tx, err := ls.db.BeginTx(ctx, nil)
-	if err != nil {
-		return fmt.Errorf("failed to begin transaction: %w", err)
-	}
-	defer func() {
-		if err != nil {
-			rollbackTx(tx, "StoreComponentDID")
-		}
-	}()
-
-	// Execute with retry logic
-	err = ls.retryOnConstraintFailure(ctx, func() error {
-		// INSERT-only query - no ON CONFLICT clause for security
-		query := `
-			INSERT INTO component_dids (
-				did, agent_did, component_type, function_name, public_key_jwk, derivation_path
-			) VALUES (?, ?, ?, ?, ?, ?)`
-
-		derivationPath := fmt.Sprintf("m/44'/0'/0'/%d", derivationIndex)
-		// For now, use empty public key - this should be passed as a parameter in the future
-		publicKeyJWK := ""
-		_, execErr := tx.ExecContext(ctx, query, componentDID, agentDID, componentType, componentName, publicKeyJWK, derivationPath)
-		if execErr != nil {
-			// Check if this is a unique constraint violation (duplicate component DID)
-			if strings.Contains(execErr.Error(), "UNIQUE constraint failed") || strings.Contains(execErr.Error(), "component_dids") {
-				logger.Logger.Warn().Msgf("Duplicate component DID entry detected: agent_did=%s, function_name=%s, component_type=%s", agentDID, componentName, componentType)
-				return &DuplicateDIDError{
-					DID:  fmt.Sprintf("component:%s/%s@%s", componentType, componentName, agentDID),
-					Type: "component",
-				}
-			}
-			// Check for foreign key constraint violations
-			if strings.Contains(execErr.Error(), "FOREIGN KEY constraint failed") {
-				return &ForeignKeyConstraintError{
-					Table:           "component_dids",
-					Column:          "agent_did",
-					ReferencedTable: "agent_dids",
-					ReferencedValue: agentDID,
-					Operation:       "INSERT",
-				}
-			}
-			return fmt.Errorf("failed to store component DID: %w", execErr)
-		}
-		return nil
-	}, 3) // Retry up to 3 times for transient errors
-
-	if err != nil {
-		return err
-	}
-
-	// Commit transaction
-	if err = tx.Commit(); err != nil {
-		return fmt.Errorf("failed to commit transaction: %w", err)
-	}
-
-	logger.Logger.Debug().Msgf("Successfully stored component DID entry: component_did=%s, agent_did=%s, type=%s", componentDID, agentDID, componentType)
-	return nil
-}
-
-func (ls *LocalStorage) GetComponentDID(ctx context.Context, componentID string) (*types.ComponentDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get component DID: %w", err)
-	}
-
-	// Use function_name as the componentID since there's no separate component_id column
-	query := `
-		SELECT function_name, did, agent_did, component_type, function_name,
-			   derivation_path, created_at
-		FROM component_dids WHERE function_name = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, componentID)
-	info := &types.ComponentDIDInfo{}
-
-	var derivationPath string
-	var createdAt sql.NullTime
-
-	err := row.Scan(&info.ComponentID, &info.ComponentDID, &info.AgentDID,
-		&info.ComponentType, &info.ComponentName, &derivationPath, &createdAt)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("component DID for %s not found", componentID)
-		}
-		return nil, fmt.Errorf("failed to get component DID: %w", err)
-	}
-
-	if createdAt.Valid {
-		info.CreatedAt = createdAt.Time
-	}
-
-	// Parse derivation index from derivation path (e.g., "m/44'/0'/0'/123" -> 123)
-	if derivationPath != "" {
-		parts := strings.Split(derivationPath, "/")
-		if len(parts) > 0 {
-			lastPart := parts[len(parts)-1]
-			if derivationIndex, parseErr := strconv.Atoi(strings.Trim(lastPart, "'")); parseErr == nil {
-				info.DerivationIndex = derivationIndex
-			}
-		}
-	}
-
-	return info, nil
-}
-
-func (ls *LocalStorage) ListComponentDIDs(ctx context.Context, agentDID string) ([]*types.ComponentDIDInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list component DIDs: %w", err)
-	}
-
-	var query string
-	var rows *sql.Rows
-	var err error
-
-	if agentDID == "" {
-		// Get all components when agentDID is empty
-		query = `
-			SELECT function_name, did, agent_did, component_type, function_name,
-				   derivation_path, created_at
-			FROM component_dids ORDER BY created_at DESC`
-		rows, err = ls.db.QueryContext(ctx, query)
-	} else {
-		// Get components for specific agent
-		query = `
-			SELECT function_name, did, agent_did, component_type, function_name,
-				   derivation_path, created_at
-			FROM component_dids WHERE agent_did = ? ORDER BY created_at DESC`
-		rows, err = ls.db.QueryContext(ctx, query, agentDID)
-	}
-	if err != nil {
-		return nil, fmt.Errorf("failed to list component DIDs: %w", err)
-	}
-	defer rows.Close()
-
-	var infos []*types.ComponentDIDInfo
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during component DID list iteration: %w", err)
-		}
-
-		info := &types.ComponentDIDInfo{}
-		var derivationPath string
-		var createdAt sql.NullTime
-
-		err := rows.Scan(&info.ComponentID, &info.ComponentDID, &info.AgentDID,
-			&info.ComponentType, &info.ComponentName, &derivationPath, &createdAt)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan component DID: %w", err)
-		}
-
-		if createdAt.Valid {
-			info.CreatedAt = createdAt.Time
-		}
-
-		// Parse derivation index from derivation path
-		if derivationPath != "" {
-			parts := strings.Split(derivationPath, "/")
-			if len(parts) > 0 {
-				lastPart := parts[len(parts)-1]
-				if derivationIndex, parseErr := strconv.Atoi(strings.Trim(lastPart, "'")); parseErr == nil {
-					info.DerivationIndex = derivationIndex
-				}
-			}
-		}
-
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-// Execution VC operations
-func (ls *LocalStorage) StoreExecutionVC(ctx context.Context, vcID, executionID, workflowID, sessionID, issuerDID, targetDID, callerDID, inputHash, outputHash, status string, vcDocument []byte, signature string, storageURI string, documentSizeBytes int64) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store execution VC: %w", err)
-	}
-
-	query := `
-		INSERT INTO execution_vcs (
-			vc_id, execution_id, workflow_id, session_id, issuer_did, target_did,
-			caller_did, vc_document, signature, storage_uri, document_size_bytes,
-			input_hash, output_hash, status, created_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-		ON CONFLICT(vc_id) DO UPDATE SET
-			status = excluded.status,
-			vc_document = excluded.vc_document,
-			signature = excluded.signature,
-			storage_uri = excluded.storage_uri,
-			document_size_bytes = excluded.document_size_bytes;`
-
-	_, err := ls.db.ExecContext(ctx, query, vcID, executionID, workflowID, sessionID, issuerDID, targetDID,
-		callerDID, vcDocument, signature, storageURI, documentSizeBytes, inputHash, outputHash, status, time.Now())
-	if err != nil {
-		return fmt.Errorf("failed to store execution VC: %w", err)
-	}
-	return nil
-}
-
-// StoreExecutionVCRecord persists an ExecutionVC including the kind
-// discriminator, parent_vc_id chain pointer, and trigger-event metadata.
-// Used by all new VC writers — the older scalar StoreExecutionVC stays for
-// backward compatibility with existing call sites.
-func (ls *LocalStorage) StoreExecutionVCRecord(ctx context.Context, vc *types.ExecutionVC) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store execution VC: %w", err)
-	}
-	if vc == nil {
-		return fmt.Errorf("execution VC is nil")
-	}
-	kind := vc.Kind
-	if kind == "" {
-		kind = types.ExecutionVCKindExecution
-	}
-	created := vc.CreatedAt
-	if created.IsZero() {
-		created = time.Now()
-	}
-
-	query := `
-		INSERT INTO execution_vcs (
-			vc_id, execution_id, workflow_id, session_id, issuer_did, target_did,
-			caller_did, vc_document, signature, storage_uri, document_size_bytes,
-			input_hash, output_hash, status, parent_vc_id, kind,
-			trigger_id, source_name, event_type, event_id, created_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-		ON CONFLICT(vc_id) DO UPDATE SET
-			status = excluded.status,
-			vc_document = excluded.vc_document,
-			signature = excluded.signature,
-			storage_uri = excluded.storage_uri,
-			document_size_bytes = excluded.document_size_bytes,
-			parent_vc_id = excluded.parent_vc_id,
-			kind = excluded.kind,
-			trigger_id = excluded.trigger_id,
-			source_name = excluded.source_name,
-			event_type = excluded.event_type,
-			event_id = excluded.event_id;`
-
-	_, err := ls.db.ExecContext(ctx, query,
-		vc.VCID, vc.ExecutionID, vc.WorkflowID, vc.SessionID, vc.IssuerDID, vc.TargetDID,
-		vc.CallerDID, []byte(vc.VCDocument), vc.Signature, vc.StorageURI, vc.DocumentSize,
-		vc.InputHash, vc.OutputHash, vc.Status, vc.ParentVCID, kind,
-		vc.TriggerID, vc.SourceName, vc.EventType, vc.EventID, created,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to store execution VC record: %w", err)
-	}
-	return nil
-}
-
-func (ls *LocalStorage) GetExecutionVC(ctx context.Context, vcID string) (*types.ExecutionVCInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get execution VC: %w", err)
-	}
-
-	query := `
-		SELECT vc_id, execution_id, workflow_id, session_id, issuer_did, target_did,
-			   caller_did, input_hash, output_hash, status, created_at, storage_uri, document_size_bytes,
-			   parent_vc_id, kind, trigger_id, source_name, event_type, event_id
-		FROM execution_vcs WHERE vc_id = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, vcID)
-	info := &types.ExecutionVCInfo{}
-
-	err := row.Scan(&info.VCID, &info.ExecutionID, &info.WorkflowID, &info.SessionID,
-		&info.IssuerDID, &info.TargetDID, &info.CallerDID, &info.InputHash,
-		&info.OutputHash, &info.Status, &info.CreatedAt, &info.StorageURI, &info.DocumentSize,
-		&info.ParentVCID, &info.Kind, &info.TriggerID, &info.SourceName, &info.EventType, &info.EventID)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("execution VC %s not found", vcID)
-		}
-		return nil, fmt.Errorf("failed to get execution VC: %w", err)
-	}
-	return info, nil
-}
-
-func buildExecutionVCFilterClauses(filters types.VCFilters) (string, []interface{}) {
-	var (
-		conditions []string
-		args       []interface{}
-	)
-
-	if filters.ExecutionID != nil {
-		conditions = append(conditions, "evc.execution_id = ?")
-		args = append(args, *filters.ExecutionID)
-	}
-	if filters.WorkflowID != nil {
-		conditions = append(conditions, "evc.workflow_id = ?")
-		args = append(args, *filters.WorkflowID)
-	}
-	if filters.SessionID != nil {
-		conditions = append(conditions, "evc.session_id = ?")
-		args = append(args, *filters.SessionID)
-	}
-	if filters.IssuerDID != nil {
-		conditions = append(conditions, "evc.issuer_did = ?")
-		args = append(args, *filters.IssuerDID)
-	}
-	if filters.TargetDID != nil {
-		conditions = append(conditions, "evc.target_did = ?")
-		args = append(args, *filters.TargetDID)
-	}
-	if filters.CallerDID != nil {
-		conditions = append(conditions, "evc.caller_did = ?")
-		args = append(args, *filters.CallerDID)
-	}
-	if filters.AgentNodeID != nil {
-		conditions = append(conditions, "COALESCE(we.agent_node_id, '') = ?")
-		args = append(args, *filters.AgentNodeID)
-	}
-	if filters.Status != nil {
-		conditions = append(conditions, "evc.status = ?")
-		args = append(args, *filters.Status)
-	}
-	if filters.CreatedAfter != nil {
-		conditions = append(conditions, "evc.created_at >= ?")
-		args = append(args, filters.CreatedAfter.UTC())
-	}
-	if filters.CreatedBefore != nil {
-		conditions = append(conditions, "evc.created_at <= ?")
-		args = append(args, filters.CreatedBefore.UTC())
-	}
-
-	if filters.Search != nil {
-		if trimmed := strings.TrimSpace(*filters.Search); trimmed != "" {
-			search := "%" + strings.ToLower(trimmed) + "%"
-			conditions = append(conditions, "("+
-				"LOWER(evc.execution_id) LIKE ? OR "+
-				"LOWER(evc.workflow_id) LIKE ? OR "+
-				"LOWER(evc.issuer_did) LIKE ? OR "+
-				"LOWER(evc.target_did) LIKE ? OR "+
-				"LOWER(evc.caller_did) LIKE ? OR "+
-				"LOWER(evc.session_id) LIKE ? OR "+
-				"LOWER(COALESCE(we.agent_node_id, '')) LIKE ? OR "+
-				"LOWER(COALESCE(we.workflow_name, '')) LIKE ?"+
-				")")
-			for i := 0; i < 8; i++ {
-				args = append(args, search)
-			}
-		}
-	}
-
-	return strings.Join(conditions, " AND "), args
-}
-
-func (ls *LocalStorage) ListExecutionVCs(ctx context.Context, filters types.VCFilters) ([]*types.ExecutionVCInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list execution VCs: %w", err)
-	}
-
-	query := `
-		SELECT evc.vc_id, evc.execution_id, evc.workflow_id, evc.session_id,
-		       evc.issuer_did, evc.target_did, evc.caller_did, evc.input_hash,
-		       evc.output_hash, evc.status, evc.created_at, evc.storage_uri,
-		       evc.document_size_bytes, we.agent_node_id, we.workflow_name,
-		       evc.parent_vc_id, evc.kind, evc.trigger_id, evc.source_name, evc.event_type, evc.event_id
-		FROM execution_vcs evc
-		LEFT JOIN workflow_executions we ON we.execution_id = evc.execution_id`
-
-	whereClause, args := buildExecutionVCFilterClauses(filters)
-	if whereClause != "" {
-		query += " WHERE " + whereClause
-	}
-
-	query += " ORDER BY evc.created_at DESC"
-
-	if filters.Limit > 0 {
-		query += fmt.Sprintf(" LIMIT %d", filters.Limit)
-	}
-
-	if filters.Offset > 0 {
-		query += fmt.Sprintf(" OFFSET %d", filters.Offset)
-	}
-
-	rows, err := ls.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list execution VCs: %w", err)
-	}
-	defer rows.Close()
-
-	var infos []*types.ExecutionVCInfo
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during execution VC list iteration: %w", err)
-		}
-
-		info := &types.ExecutionVCInfo{}
-		err := rows.Scan(&info.VCID, &info.ExecutionID, &info.WorkflowID, &info.SessionID,
-			&info.IssuerDID, &info.TargetDID, &info.CallerDID, &info.InputHash,
-			&info.OutputHash, &info.Status, &info.CreatedAt, &info.StorageURI,
-			&info.DocumentSize, &info.AgentNodeID, &info.WorkflowName,
-			&info.ParentVCID, &info.Kind, &info.TriggerID, &info.SourceName, &info.EventType, &info.EventID)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan execution VC: %w", err)
-		}
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-func (ls *LocalStorage) ListWorkflowVCStatusSummaries(ctx context.Context, workflowIDs []string) ([]*types.WorkflowVCStatusAggregation, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during workflow VC status summary query: %w", err)
-	}
-
-	if len(workflowIDs) == 0 {
-		return []*types.WorkflowVCStatusAggregation{}, nil
-	}
-
-	placeholders := make([]string, len(workflowIDs))
-	for i := range workflowIDs {
-		placeholders[i] = "?"
-	}
-
-	query := fmt.Sprintf(`
-		SELECT workflow_id,
-		       COUNT(*) AS vc_count,
-		       SUM(CASE WHEN status = ? THEN 1 ELSE 0 END) AS verified_count,
-		       SUM(CASE WHEN status = ? OR status = ? THEN 1 ELSE 0 END) AS failed_count,
-		       MAX(created_at) AS last_created_at
-		FROM execution_vcs
-		WHERE workflow_id IN (%s)
-		GROUP BY workflow_id
-	`, strings.Join(placeholders, ","))
-
-	args := []interface{}{
-		string(types.ExecutionStatusSucceeded),
-		string(types.ExecutionStatusFailed),
-		string(types.ExecutionStatusTimeout),
-	}
-	for _, id := range workflowIDs {
-		args = append(args, id)
-	}
-
-	rows, err := ls.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query workflow VC status summaries: %w", err)
-	}
-	defer rows.Close()
-
-	var summaries []*types.WorkflowVCStatusAggregation
-	for rows.Next() {
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during workflow VC status iteration: %w", err)
-		}
-
-		var lastCreated sql.NullTime
-		summary := &types.WorkflowVCStatusAggregation{}
-		if err := rows.Scan(
-			&summary.WorkflowID,
-			&summary.VCCount,
-			&summary.VerifiedCount,
-			&summary.FailedCount,
-			&lastCreated,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan workflow VC status summary: %w", err)
-		}
-
-		if lastCreated.Valid {
-			summary.LastCreatedAt = &lastCreated.Time
-		}
-
-		summaries = append(summaries, summary)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("workflow VC status summary rows error: %w", err)
-	}
-
-	return summaries, nil
-}
-
-func (ls *LocalStorage) CountExecutionVCs(ctx context.Context, filters types.VCFilters) (int, error) {
-	if err := ctx.Err(); err != nil {
-		return 0, fmt.Errorf("context cancelled during count execution VCs: %w", err)
-	}
-
-	query := `
-		SELECT COUNT(*)
-		FROM execution_vcs evc
-		LEFT JOIN workflow_executions we ON we.execution_id = evc.execution_id`
-
-	whereClause, args := buildExecutionVCFilterClauses(filters)
-	if whereClause != "" {
-		query += " WHERE " + whereClause
-	}
-
-	var total int
-	if err := ls.db.QueryRowContext(ctx, query, args...).Scan(&total); err != nil {
-		return 0, fmt.Errorf("failed to count execution VCs: %w", err)
-	}
-	return total, nil
-}
-
-// Workflow VC operations
-func (ls *LocalStorage) StoreWorkflowVC(ctx context.Context, workflowVCID, workflowID, sessionID string, componentVCIDs []string, status string, startTime, endTime *time.Time, totalSteps, completedSteps int, storageURI string, documentSizeBytes int64) error {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store workflow VC: %w", err)
-	}
-
-	componentVCIDsJSON, err := json.Marshal(componentVCIDs)
-	if err != nil {
-		return fmt.Errorf("failed to marshal component VC IDs: %w", err)
-	}
-
-	query := `
-		INSERT INTO workflow_vcs (
-			workflow_vc_id, workflow_id, session_id, component_vc_ids, status,
-			start_time, end_time, total_steps, completed_steps, storage_uri, document_size_bytes
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-		ON CONFLICT(workflow_vc_id) DO UPDATE SET
-			component_vc_ids = excluded.component_vc_ids,
-			status = excluded.status,
-			end_time = excluded.end_time,
-			completed_steps = excluded.completed_steps,
-			storage_uri = excluded.storage_uri,
-			document_size_bytes = excluded.document_size_bytes;`
-
-	_, err = ls.db.ExecContext(ctx, query, workflowVCID, workflowID, sessionID, componentVCIDsJSON, status,
-		startTime, endTime, totalSteps, completedSteps, storageURI, documentSizeBytes)
-	if err != nil {
-		return fmt.Errorf("failed to store workflow VC: %w", err)
-	}
-	return nil
-}
-
-func (ls *LocalStorage) GetWorkflowVC(ctx context.Context, workflowVCID string) (*types.WorkflowVCInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get workflow VC: %w", err)
-	}
-
-	query := `
-		SELECT workflow_vc_id, workflow_id, session_id, component_vc_ids, status,
-			   start_time, end_time, total_steps, completed_steps, storage_uri, document_size_bytes
-		FROM workflow_vcs WHERE workflow_vc_id = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, workflowVCID)
-	info := &types.WorkflowVCInfo{}
-	var componentVCIDsJSON []byte
-
-	err := row.Scan(&info.WorkflowVCID, &info.WorkflowID, &info.SessionID, &componentVCIDsJSON,
-		&info.Status, &info.StartTime, &info.EndTime, &info.TotalSteps, &info.CompletedSteps, &info.StorageURI, &info.DocumentSize)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("workflow VC %s not found", workflowVCID)
-		}
-		return nil, fmt.Errorf("failed to get workflow VC: %w", err)
-	}
-
-	if len(componentVCIDsJSON) > 0 {
-		if err := json.Unmarshal(componentVCIDsJSON, &info.ComponentVCIDs); err != nil {
-			return nil, fmt.Errorf("failed to unmarshal component VC IDs: %w", err)
-		}
-	}
-
-	return info, nil
-}
-
-func (ls *LocalStorage) ListWorkflowVCs(ctx context.Context, workflowID string) ([]*types.WorkflowVCInfo, error) {
-	// Check context cancellation early
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list workflow VCs: %w", err)
-	}
-
-	var query string
-	var args []interface{}
-
-	if workflowID == "" {
-		// Get all workflow VCs
-		query = `
-			SELECT workflow_vc_id, workflow_id, session_id, component_vc_ids, status,
-				   start_time, end_time, total_steps, completed_steps, storage_uri, document_size_bytes
-			FROM workflow_vcs ORDER BY start_time DESC`
-	} else {
-		// Get workflow VCs for specific workflow
-		query = `
-			SELECT workflow_vc_id, workflow_id, session_id, component_vc_ids, status,
-				   start_time, end_time, total_steps, completed_steps, storage_uri, document_size_bytes
-			FROM workflow_vcs WHERE workflow_id = ? ORDER BY start_time DESC`
-		args = append(args, workflowID)
-	}
-
-	rows, err := ls.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list workflow VCs: %w", err)
-	}
-	defer rows.Close()
-
-	var infos []*types.WorkflowVCInfo
-	for rows.Next() {
-		// Check context cancellation during iteration
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during workflow VC list iteration: %w", err)
-		}
-
-		info := &types.WorkflowVCInfo{}
-		var componentVCIDsJSON []byte
-
-		err := rows.Scan(&info.WorkflowVCID, &info.WorkflowID, &info.SessionID, &componentVCIDsJSON,
-			&info.Status, &info.StartTime, &info.EndTime, &info.TotalSteps, &info.CompletedSteps, &info.StorageURI, &info.DocumentSize)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan workflow VC: %w", err)
-		}
-
-		if len(componentVCIDsJSON) > 0 {
-			if err := json.Unmarshal(componentVCIDsJSON, &info.ComponentVCIDs); err != nil {
-				return nil, fmt.Errorf("failed to unmarshal component VC IDs: %w", err)
-			}
-		}
-
-		infos = append(infos, info)
-	}
-	return infos, nil
-}
-
-// GetFullExecutionVC retrieves the full execution VC including the VC document and signature
-func (ls *LocalStorage) GetFullExecutionVC(vcID string) (json.RawMessage, string, error) {
-	query := `
-		SELECT vc_document, signature
-		FROM execution_vcs WHERE vc_id = ?`
-
-	row := ls.db.QueryRow(query, vcID)
-
-	var vcDocument string
-	var signature string
-
-	err := row.Scan(&vcDocument, &signature)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, "", fmt.Errorf("execution VC %s not found", vcID)
-		}
-		return nil, "", fmt.Errorf("failed to get full execution VC: %w", err)
-	}
-
-	return json.RawMessage(vcDocument), signature, nil
-}
-
-// TransactionalStorage methods (not fully implemented for local storage yet)
-func (ls *LocalStorage) BeginTransaction() (Transaction, error) {
-	return nil, fmt.Errorf("transactions not fully implemented for LocalStorage")
-}
-
-// StoreWorkflowExecutionEvent inserts an immutable execution event into SQLite.
-func (ls *LocalStorage) StoreWorkflowExecutionEvent(ctx context.Context, event *types.WorkflowExecutionEvent) error {
-	if event == nil {
-		return fmt.Errorf("workflow execution event is nil")
-	}
-
-	// Use retry logic for database lock errors
-	return ls.retryDatabaseOperation(ctx, event.ExecutionID, func() error {
-		tx, err := ls.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("failed to begin transaction: %w", err)
-		}
-		defer rollbackTx(tx, "StoreWorkflowExecutionEvent:"+event.ExecutionID)
-
-		if err := ls.storeWorkflowExecutionEventTx(ctx, tx, event); err != nil {
-			return err
-		}
-
-		if err := tx.Commit(); err != nil {
-			return fmt.Errorf("failed to commit event transaction: %w", err)
-		}
-
-		eventCopy := *event
-		ls.workflowExecutionEventBus.Publish(&eventCopy)
-
-		return nil
-	})
-}
-
-// storeWorkflowExecutionEventTx inserts an execution event within an existing transaction.
-// This allows atomic operations where the event storage and execution update happen together.
-func (ls *LocalStorage) storeWorkflowExecutionEventTx(ctx context.Context, tx DBTX, event *types.WorkflowExecutionEvent) error {
-	payload := string(event.Payload)
-	if len(event.Payload) == 0 {
-		payload = "{}"
-	}
-
-	// Persist recorded_at explicitly: the GORM model's autoCreateTime does
-	// not apply to this raw INSERT, and a NULL recorded_at breaks listing.
-	if event.RecordedAt.IsZero() {
-		event.RecordedAt = time.Now().UTC()
-	}
-
-	query := `
-		INSERT INTO workflow_execution_events (
-			execution_id, workflow_id, run_id, parent_execution_id, sequence, previous_sequence,
-			event_type, status, status_reason, payload, emitted_at, recorded_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-
-	result, err := tx.ExecContext(ctx, query,
-		event.ExecutionID,
-		event.WorkflowID,
-		event.RunID,
-		event.ParentExecutionID,
-		event.Sequence,
-		event.PreviousSequence,
-		event.EventType,
-		event.Status,
-		event.StatusReason,
-		payload,
-		event.EmittedAt,
-		event.RecordedAt,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to insert workflow execution event: %w", err)
-	}
-
-	if id, err := result.LastInsertId(); err == nil {
-		event.EventID = id
-	}
-
-	return nil
-}
-
-// ListWorkflowExecutionEvents retrieves execution events ordered by sequence.
-func (ls *LocalStorage) ListWorkflowExecutionEvents(ctx context.Context, executionID string, afterSeq *int64, limit int) ([]*types.WorkflowExecutionEvent, error) {
-	query := `
-		SELECT event_id, execution_id, workflow_id, run_id, parent_execution_id, sequence, previous_sequence,
-		       event_type, status, status_reason, payload, emitted_at, recorded_at
-		FROM workflow_execution_events
-		WHERE execution_id = ?`
-	args := []interface{}{executionID}
-
-	if afterSeq != nil {
-		query += " AND sequence > ?"
-		args = append(args, *afterSeq)
-	}
-
-	query += " ORDER BY sequence ASC"
-	if limit > 0 {
-		query += fmt.Sprintf(" LIMIT %d", limit)
-	}
-
-	rows, err := ls.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query workflow execution events: %w", err)
-	}
-	defer rows.Close()
-
-	var events []*types.WorkflowExecutionEvent
-	for rows.Next() {
-		evt := &types.WorkflowExecutionEvent{}
-		var runID sql.NullString
-		var parentID sql.NullString
-		var status sql.NullString
-		var statusReason sql.NullString
-		var payload sql.NullString
-		var recordedAt sql.NullTime
-
-		if err := rows.Scan(
-			&evt.EventID,
-			&evt.ExecutionID,
-			&evt.WorkflowID,
-			&runID,
-			&parentID,
-			&evt.Sequence,
-			&evt.PreviousSequence,
-			&evt.EventType,
-			&status,
-			&statusReason,
-			&payload,
-			&evt.EmittedAt,
-			&recordedAt,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan workflow execution event: %w", err)
-		}
-
-		// Rows written before recorded_at was persisted have NULL here;
-		// fall back to emitted_at rather than failing the whole listing.
-		if recordedAt.Valid {
-			evt.RecordedAt = recordedAt.Time
-		} else {
-			evt.RecordedAt = evt.EmittedAt
-		}
-
-		if runID.Valid {
-			evt.RunID = &runID.String
-		}
-		if parentID.Valid {
-			evt.ParentExecutionID = &parentID.String
-		}
-		if status.Valid {
-			value := status.String
-			evt.Status = &value
-		}
-		if statusReason.Valid {
-			value := statusReason.String
-			evt.StatusReason = &value
-		}
-		if payload.Valid {
-			evt.Payload = json.RawMessage(payload.String)
-		} else {
-			evt.Payload = json.RawMessage("{}")
-		}
-
-		events = append(events, evt)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating workflow execution events: %w", err)
-	}
-
-	return events, nil
-}
-
-// StoreExecutionLogEntry inserts a structured execution log entry and publishes it to subscribers.
-func (ls *LocalStorage) StoreExecutionLogEntry(ctx context.Context, entry *types.ExecutionLogEntry) error {
-	if entry == nil {
-		return fmt.Errorf("execution log entry is nil")
-	}
-	if strings.TrimSpace(entry.ExecutionID) == "" {
-		return fmt.Errorf("execution_id is required")
-	}
-	if strings.TrimSpace(entry.WorkflowID) == "" {
-		entry.WorkflowID = entry.ExecutionID
-	}
-	if strings.TrimSpace(entry.Level) == "" {
-		entry.Level = "info"
-	}
-	if strings.TrimSpace(entry.Source) == "" {
-		entry.Source = "sdk.logger"
-	}
-	if entry.EmittedAt.IsZero() {
-		entry.EmittedAt = time.Now().UTC()
-	}
-	ls.mu.Lock()
-	defer ls.mu.Unlock()
-
-	return ls.retryDatabaseOperation(ctx, entry.ExecutionID, func() error {
-		tx, err := ls.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("failed to begin execution log transaction: %w", err)
-		}
-		defer rollbackTx(tx, "StoreExecutionLogEntry:"+entry.ExecutionID)
-
-		if err := ls.storeExecutionLogEntryTx(ctx, tx, entry); err != nil {
-			return err
-		}
-
-		if err := tx.Commit(); err != nil {
-			return fmt.Errorf("failed to commit execution log transaction: %w", err)
-		}
-
-		entryCopy := *entry
-		ls.executionLogEventBus.Publish(&entryCopy)
-		return nil
-	})
-}
-
-// StoreExecutionLogEntries atomically stores a batch of structured execution logs for one execution.
-func (ls *LocalStorage) StoreExecutionLogEntries(ctx context.Context, executionID string, entries []*types.ExecutionLogEntry) error {
-	if len(entries) == 0 {
-		return nil
-	}
-
-	ls.mu.Lock()
-	defer ls.mu.Unlock()
-
-	return ls.retryDatabaseOperation(ctx, executionID, func() error {
-		tx, err := ls.db.BeginTx(ctx, nil)
-		if err != nil {
-			return fmt.Errorf("failed to begin execution log batch transaction: %w", err)
-		}
-		defer rollbackTx(tx, "StoreExecutionLogEntries:"+executionID)
-
-		for _, entry := range entries {
-			if entry == nil {
-				continue
-			}
-			if err := ls.storeExecutionLogEntryTx(ctx, tx, entry); err != nil {
-				return err
-			}
-		}
-
-		if err := tx.Commit(); err != nil {
-			return fmt.Errorf("failed to commit execution log batch transaction: %w", err)
-		}
-
-		for _, entry := range entries {
-			if entry == nil {
-				continue
-			}
-			entryCopy := *entry
-			ls.executionLogEventBus.Publish(&entryCopy)
-		}
-		return nil
-	})
-}
-
-func (ls *LocalStorage) storeExecutionLogEntryTx(ctx context.Context, tx DBTX, entry *types.ExecutionLogEntry) error {
-	var nextSeq int64 = 1
-	if err := tx.QueryRowContext(ctx,
-		`SELECT COALESCE(MAX(sequence), 0) + 1 FROM execution_logs WHERE execution_id = ?`,
-		entry.ExecutionID,
-	).Scan(&nextSeq); err != nil {
-		return fmt.Errorf("failed to compute execution log sequence: %w", err)
-	}
-	entry.Sequence = nextSeq
-
-	attributes := "{}"
-	if len(entry.Attributes) > 0 {
-		attributes = string(entry.Attributes)
-	}
-
-	// Persist recorded_at explicitly: the GORM model's autoCreateTime does
-	// not apply to this raw INSERT, so omitting the column stores NULL and
-	// every read silently takes the legacy emitted_at fallback.
-	if entry.RecordedAt.IsZero() {
-		entry.RecordedAt = time.Now().UTC()
-	}
-
-	query := `
-		INSERT INTO execution_logs (
-			execution_id, workflow_id, run_id, root_workflow_id, parent_execution_id, sequence,
-			agent_node_id, reasoner_id, level, source, event_type, message, attributes,
-			system_generated, sdk_language, attempt, span_id, step_id, error_category, emitted_at,
-			recorded_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-
-	result, err := tx.ExecContext(ctx, query,
-		entry.ExecutionID,
-		entry.WorkflowID,
-		entry.RunID,
-		entry.RootWorkflowID,
-		entry.ParentExecutionID,
-		entry.Sequence,
-		entry.AgentNodeID,
-		entry.ReasonerID,
-		entry.Level,
-		entry.Source,
-		entry.EventType,
-		entry.Message,
-		attributes,
-		entry.SystemGenerated,
-		entry.SDKLanguage,
-		entry.Attempt,
-		entry.SpanID,
-		entry.StepID,
-		entry.ErrorCategory,
-		entry.EmittedAt,
-		entry.RecordedAt,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to insert execution log entry: %w", err)
-	}
-	if id, err := result.LastInsertId(); err == nil {
-		entry.EventID = id
-	}
-	return nil
-}
-
-// ListExecutionLogEntries retrieves structured execution logs ordered by sequence.
-func (ls *LocalStorage) ListExecutionLogEntries(ctx context.Context, executionID string, afterSeq *int64, limit int, levels []string, nodeIDs []string, sources []string, query string) ([]*types.ExecutionLogEntry, error) {
-	baseQuery := `
-		SELECT event_id, execution_id, workflow_id, run_id, root_workflow_id, parent_execution_id, sequence,
-		       agent_node_id, reasoner_id, level, source, event_type, message, attributes, system_generated,
-		       sdk_language, attempt, span_id, step_id, error_category, emitted_at, recorded_at
-		FROM execution_logs
-		WHERE execution_id = ?`
-	args := []interface{}{executionID}
-
-	appendIn := func(column string, values []string) {
-		if len(values) == 0 {
-			return
-		}
-		holders := make([]string, 0, len(values))
-		for _, value := range values {
-			if strings.TrimSpace(value) == "" {
-				continue
-			}
-			holders = append(holders, "?")
-			args = append(args, value)
-		}
-		if len(holders) > 0 {
-			baseQuery += " AND " + column + " IN (" + strings.Join(holders, ",") + ")"
-		}
-	}
-
-	if afterSeq != nil {
-		baseQuery += " AND sequence > ?"
-		args = append(args, *afterSeq)
-	}
-	appendIn("level", levels)
-	appendIn("agent_node_id", nodeIDs)
-	appendIn("source", sources)
-	if trimmed := strings.TrimSpace(query); trimmed != "" {
-		baseQuery += " AND (message LIKE ? OR attributes LIKE ?)"
-		like := "%" + trimmed + "%"
-		args = append(args, like, like)
-	}
-
-	descendingTail := afterSeq == nil && limit > 0
-	if descendingTail {
-		baseQuery += " ORDER BY sequence DESC"
-		baseQuery += fmt.Sprintf(" LIMIT %d", limit)
-	} else {
-		baseQuery += " ORDER BY sequence ASC"
-		if limit > 0 {
-			baseQuery += fmt.Sprintf(" LIMIT %d", limit)
-		}
-	}
-
-	rows, err := ls.db.QueryContext(ctx, baseQuery, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query execution logs: %w", err)
-	}
-	defer rows.Close()
-
-	var entries []*types.ExecutionLogEntry
-	for rows.Next() {
-		entry := &types.ExecutionLogEntry{}
-		var runID, rootWorkflowID, parentExecutionID, reasonerID, eventType, sdkLanguage, spanID, stepID, errorCategory sql.NullString
-		var attributes sql.NullString
-		var attempt sql.NullInt64
-		var emittedAt sql.NullTime
-		var recordedAt sql.NullTime
-		if err := rows.Scan(
-			&entry.EventID,
-			&entry.ExecutionID,
-			&entry.WorkflowID,
-			&runID,
-			&rootWorkflowID,
-			&parentExecutionID,
-			&entry.Sequence,
-			&entry.AgentNodeID,
-			&reasonerID,
-			&entry.Level,
-			&entry.Source,
-			&eventType,
-			&entry.Message,
-			&attributes,
-			&entry.SystemGenerated,
-			&sdkLanguage,
-			&attempt,
-			&spanID,
-			&stepID,
-			&errorCategory,
-			&emittedAt,
-			&recordedAt,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan execution log entry: %w", err)
-		}
-
-		if runID.Valid {
-			entry.RunID = &runID.String
-		}
-		if rootWorkflowID.Valid {
-			entry.RootWorkflowID = &rootWorkflowID.String
-		}
-		if parentExecutionID.Valid {
-			entry.ParentExecutionID = &parentExecutionID.String
-		}
-		if reasonerID.Valid {
-			entry.ReasonerID = &reasonerID.String
-		}
-		if eventType.Valid {
-			entry.EventType = &eventType.String
-		}
-		if sdkLanguage.Valid {
-			entry.SDKLanguage = &sdkLanguage.String
-		}
-		if spanID.Valid {
-			entry.SpanID = &spanID.String
-		}
-		if stepID.Valid {
-			entry.StepID = &stepID.String
-		}
-		if errorCategory.Valid {
-			entry.ErrorCategory = &errorCategory.String
-		}
-		if attempt.Valid {
-			value := int(attempt.Int64)
-			entry.Attempt = &value
-		}
-		if emittedAt.Valid {
-			entry.EmittedAt = emittedAt.Time
-		}
-		if recordedAt.Valid {
-			entry.RecordedAt = recordedAt.Time
-		} else {
-			entry.RecordedAt = entry.EmittedAt
-		}
-		if attributes.Valid {
-			entry.Attributes = json.RawMessage(attributes.String)
-		} else {
-			entry.Attributes = json.RawMessage("{}")
-		}
-
-		entries = append(entries, entry)
-	}
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating execution logs: %w", err)
-	}
-	if descendingTail {
-		for left, right := 0, len(entries)-1; left < right; left, right = left+1, right-1 {
-			entries[left], entries[right] = entries[right], entries[left]
-		}
-	}
-	return entries, nil
-}
-
-// PruneExecutionLogEntries trims old or excessive execution logs for a single execution.
-func (ls *LocalStorage) PruneExecutionLogEntries(ctx context.Context, executionID string, maxEntries int, olderThan time.Time) error {
-	if strings.TrimSpace(executionID) == "" {
-		return nil
-	}
-	if !olderThan.IsZero() {
-		if _, err := ls.db.ExecContext(ctx,
-			`DELETE FROM execution_logs WHERE execution_id = ? AND emitted_at < ?`,
-			executionID, olderThan,
-		); err != nil {
-			return fmt.Errorf("failed to prune execution logs by age: %w", err)
-		}
-	}
-	if maxEntries > 0 {
-		if _, err := ls.db.ExecContext(ctx, `
-			DELETE FROM execution_logs
-			WHERE execution_id = ?
-			  AND event_id NOT IN (
-			    SELECT event_id FROM execution_logs
-			    WHERE execution_id = ?
-			    ORDER BY sequence DESC
-			    LIMIT ?
-			  )`,
-			executionID, executionID, maxEntries,
-		); err != nil {
-			return fmt.Errorf("failed to prune execution logs by count: %w", err)
-		}
-	}
-	return nil
-}
-
-// StoreExecutionWebhookEvent records webhook delivery attempts for SQLite deployments.
-func (ls *LocalStorage) StoreExecutionWebhookEvent(ctx context.Context, event *types.ExecutionWebhookEvent) error {
-	if event == nil {
-		return fmt.Errorf("execution webhook event is nil")
-	}
-
-	payload := interface{}(nil)
-	if len(event.Payload) > 0 {
-		payload = string(event.Payload)
-	}
-
-	query := `
-		INSERT INTO execution_webhook_events (
-			execution_id, event_type, status, http_status, payload, response_body, error_message, created_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)`
-
-	_, err := ls.db.ExecContext(ctx, query,
-		event.ExecutionID,
-		event.EventType,
-		event.Status,
-		event.HTTPStatus,
-		payload,
-		event.ResponseBody,
-		event.ErrorMessage,
-	)
-	return err
-}
-
-// ListExecutionWebhookEvents returns webhook attempts ordered by creation time.
-func (ls *LocalStorage) ListExecutionWebhookEvents(ctx context.Context, executionID string) ([]*types.ExecutionWebhookEvent, error) {
-	query := `
-		SELECT id, execution_id, event_type, status, http_status, payload, response_body, error_message, created_at
-		FROM execution_webhook_events
-		WHERE execution_id = ?
-		ORDER BY created_at ASC, id ASC`
-
-	rows, err := ls.db.QueryContext(ctx, query, executionID)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query execution webhook events: %w", err)
-	}
-	defer rows.Close()
-
-	var events []*types.ExecutionWebhookEvent
-	for rows.Next() {
-		evt := &types.ExecutionWebhookEvent{}
-		var payload sql.NullString
-		var response sql.NullString
-		var errMsg sql.NullString
-		var status sql.NullInt64
-
-		if err := rows.Scan(
-			&evt.ID,
-			&evt.ExecutionID,
-			&evt.EventType,
-			&evt.Status,
-			&status,
-			&payload,
-			&response,
-			&errMsg,
-			&evt.CreatedAt,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan execution webhook event: %w", err)
-		}
-
-		if status.Valid {
-			s := int(status.Int64)
-			evt.HTTPStatus = &s
-		}
-		if payload.Valid {
-			evt.Payload = json.RawMessage(payload.String)
-		} else {
-			evt.Payload = json.RawMessage("{}")
-		}
-		if response.Valid {
-			value := response.String
-			evt.ResponseBody = &value
-		}
-		if errMsg.Valid {
-			value := errMsg.String
-			evt.ErrorMessage = &value
-		}
-
-		events = append(events, evt)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating execution webhook events: %w", err)
-	}
-
-	return events, nil
-}
-
-// ListExecutionWebhookEventsBatch fetches webhook events for multiple executions in a single query.
-func (ls *LocalStorage) ListExecutionWebhookEventsBatch(ctx context.Context, executionIDs []string) (map[string][]*types.ExecutionWebhookEvent, error) {
-	results := make(map[string][]*types.ExecutionWebhookEvent)
-	if len(executionIDs) == 0 {
-		return results, nil
-	}
-
-	unique := make([]string, 0, len(executionIDs))
-	seen := make(map[string]struct{}, len(executionIDs))
-	for _, id := range executionIDs {
-		trimmed := strings.TrimSpace(id)
-		if trimmed == "" {
-			continue
-		}
-		if _, ok := seen[trimmed]; ok {
-			continue
-		}
-		seen[trimmed] = struct{}{}
-		unique = append(unique, trimmed)
-	}
-	if len(unique) == 0 {
-		return results, nil
-	}
-
-	placeholders := make([]string, len(unique))
-	args := make([]interface{}, len(unique))
-	for i, id := range unique {
-		placeholders[i] = "?"
-		args[i] = id
-	}
-
-	query := fmt.Sprintf(`
-		SELECT execution_id, id, event_type, status, http_status, payload, response_body, error_message, created_at
-		FROM execution_webhook_events
-		WHERE execution_id IN (%s)
-		ORDER BY execution_id ASC, created_at ASC, id ASC`, strings.Join(placeholders, ","))
-
-	rows, err := ls.db.QueryContext(ctx, query, args...)
-	if err != nil {
-		return nil, fmt.Errorf("failed to query batch webhook events: %w", err)
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		evt := &types.ExecutionWebhookEvent{}
-		var payload sql.NullString
-		var response sql.NullString
-		var errMsg sql.NullString
-		var status sql.NullInt64
-		if err := rows.Scan(
-			&evt.ExecutionID,
-			&evt.ID,
-			&evt.EventType,
-			&evt.Status,
-			&status,
-			&payload,
-			&response,
-			&errMsg,
-			&evt.CreatedAt,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan batch webhook event: %w", err)
-		}
-
-		if status.Valid {
-			s := int(status.Int64)
-			evt.HTTPStatus = &s
-		}
-		if payload.Valid {
-			evt.Payload = json.RawMessage(payload.String)
-		} else {
-			evt.Payload = json.RawMessage("{}")
-		}
-		if response.Valid {
-			value := response.String
-			evt.ResponseBody = &value
-		}
-		if errMsg.Valid {
-			value := errMsg.String
-			evt.ErrorMessage = &value
-		}
-
-		results[evt.ExecutionID] = append(results[evt.ExecutionID], evt)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating batch webhook events: %w", err)
-	}
-
-	return results, nil
-}
-
-// =============================================================================
-// DID Document Operations (did:web Resolution)
-// =============================================================================
-
-// StoreDIDDocument stores a DID document record.
-func (ls *LocalStorage) StoreDIDDocument(ctx context.Context, record *types.DIDDocumentRecord) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store DID document: %w", err)
-	}
-
-	query := `
-		INSERT INTO did_documents (
-			did, agent_id, did_document, public_key_jwk, revoked_at, created_at, updated_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?)
-		ON CONFLICT(did) DO UPDATE SET
-			agent_id = excluded.agent_id,
-			did_document = excluded.did_document,
-			public_key_jwk = excluded.public_key_jwk,
-			updated_at = excluded.updated_at`
-
-	_, err := ls.db.ExecContext(ctx, query,
-		record.DID, record.AgentID, record.DIDDocument, record.PublicKeyJWK,
-		record.RevokedAt, record.CreatedAt, record.UpdatedAt,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to store DID document: %w", err)
-	}
-
-	return nil
-}
-
-// GetDIDDocument retrieves a DID document by its DID.
-func (ls *LocalStorage) GetDIDDocument(ctx context.Context, did string) (*types.DIDDocumentRecord, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get DID document: %w", err)
-	}
-
-	query := `
-		SELECT did, agent_id, did_document, public_key_jwk, revoked_at, created_at, updated_at
-		FROM did_documents WHERE did = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, did)
-
-	record := &types.DIDDocumentRecord{}
-	var revokedAt sql.NullTime
-
-	err := row.Scan(
-		&record.DID, &record.AgentID, &record.DIDDocument, &record.PublicKeyJWK,
-		&revokedAt, &record.CreatedAt, &record.UpdatedAt,
-	)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("DID document not found: %s", did)
-		}
-		return nil, fmt.Errorf("failed to get DID document: %w", err)
-	}
-
-	if revokedAt.Valid {
-		record.RevokedAt = &revokedAt.Time
-	}
-
-	return record, nil
-}
-
-// GetDIDDocumentByAgentID retrieves a DID document by agent ID.
-func (ls *LocalStorage) GetDIDDocumentByAgentID(ctx context.Context, agentID string) (*types.DIDDocumentRecord, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get DID document by agent ID: %w", err)
-	}
-
-	query := `
-		SELECT did, agent_id, did_document, public_key_jwk, revoked_at, created_at, updated_at
-		FROM did_documents WHERE agent_id = ? AND revoked_at IS NULL
-		ORDER BY created_at DESC LIMIT 1`
-
-	row := ls.db.QueryRowContext(ctx, query, agentID)
-
-	record := &types.DIDDocumentRecord{}
-	var revokedAt sql.NullTime
-
-	err := row.Scan(
-		&record.DID, &record.AgentID, &record.DIDDocument, &record.PublicKeyJWK,
-		&revokedAt, &record.CreatedAt, &record.UpdatedAt,
-	)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("DID document not found for agent: %s", agentID)
-		}
-		return nil, fmt.Errorf("failed to get DID document by agent ID: %w", err)
-	}
-
-	if revokedAt.Valid {
-		record.RevokedAt = &revokedAt.Time
-	}
-
-	return record, nil
-}
-
-// RevokeDIDDocument revokes a DID document by setting its revoked_at timestamp.
-func (ls *LocalStorage) RevokeDIDDocument(ctx context.Context, did string) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during revoke DID document: %w", err)
-	}
-
-	query := `UPDATE did_documents SET revoked_at = ?, updated_at = ? WHERE did = ?`
-
-	now := time.Now()
-	result, err := ls.db.ExecContext(ctx, query, now, now, did)
-	if err != nil {
-		return fmt.Errorf("failed to revoke DID document: %w", err)
-	}
-
-	rows, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to get rows affected: %w", err)
-	}
-	if rows == 0 {
-		return fmt.Errorf("DID document not found: %s", did)
-	}
-
-	return nil
-}
-
-// ListDIDDocuments lists all DID documents.
-func (ls *LocalStorage) ListDIDDocuments(ctx context.Context) ([]*types.DIDDocumentRecord, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list DID documents: %w", err)
-	}
-
-	query := `
-		SELECT did, agent_id, did_document, public_key_jwk, revoked_at, created_at, updated_at
-		FROM did_documents ORDER BY created_at DESC`
-
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list DID documents: %w", err)
-	}
-	defer rows.Close()
-
-	var records []*types.DIDDocumentRecord
-	for rows.Next() {
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during scan: %w", err)
-		}
-
-		record := &types.DIDDocumentRecord{}
-		var revokedAt sql.NullTime
-
-		err := rows.Scan(
-			&record.DID, &record.AgentID, &record.DIDDocument, &record.PublicKeyJWK,
-			&revokedAt, &record.CreatedAt, &record.UpdatedAt,
-		)
-		if err != nil {
-			return nil, fmt.Errorf("failed to scan DID document: %w", err)
-		}
-
-		if revokedAt.Valid {
-			record.RevokedAt = &revokedAt.Time
-		}
-
-		records = append(records, record)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating DID documents: %w", err)
-	}
-
-	return records, nil
-}
-
-// ListAgentsByLifecycleStatus lists agents filtered by lifecycle status.
-func (ls *LocalStorage) ListAgentsByLifecycleStatus(ctx context.Context, status types.AgentLifecycleStatus) ([]*types.AgentNode, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list agents by lifecycle status: %w", err)
-	}
-
-	query := `
-		SELECT
-			id, version, group_id, team_id, base_url, traffic_weight, deployment_type, invocation_url, reasoners, skills,
-			communication_config, health_status, lifecycle_status, last_heartbeat,
-			registered_at, features, metadata, proposed_tags, approved_tags, COALESCE(instance_id, '')
-		FROM agent_nodes WHERE lifecycle_status = ? ORDER BY registered_at DESC`
-
-	rows, err := ls.db.QueryContext(ctx, query, string(status))
-	if err != nil {
-		return nil, fmt.Errorf("failed to list agents by lifecycle status: %w", err)
-	}
-	defer rows.Close()
-
-	return ls.scanAgentNodes(ctx, rows)
-}
-
-// reconstructAgentLevelTags ensures agent-level ProposedTags and ApprovedTags
-// are populated. If the dedicated DB columns were empty (e.g., on older records),
-// it reconstructs them from per-reasoner/per-skill fields as a fallback.
-func reconstructAgentLevelTags(agent *types.AgentNode) {
-	// Only reconstruct if DB columns were empty
-	if len(agent.ApprovedTags) == 0 {
-		seen := make(map[string]struct{})
-		for _, r := range agent.Reasoners {
-			for _, t := range r.ApprovedTags {
-				if _, exists := seen[t]; !exists {
-					seen[t] = struct{}{}
-					agent.ApprovedTags = append(agent.ApprovedTags, t)
-				}
-			}
-		}
-		for _, sk := range agent.Skills {
-			for _, t := range sk.ApprovedTags {
-				if _, exists := seen[t]; !exists {
-					seen[t] = struct{}{}
-					agent.ApprovedTags = append(agent.ApprovedTags, t)
-				}
-			}
-		}
-		types.HydrateAgentSessions(agent)
-		for _, session := range agent.Sessions {
-			for _, t := range session.ApprovedTags {
-				if _, exists := seen[t]; !exists {
-					seen[t] = struct{}{}
-					agent.ApprovedTags = append(agent.ApprovedTags, t)
-				}
-			}
-		}
-	}
-
-	if len(agent.ProposedTags) == 0 {
-		proposedSeen := make(map[string]struct{})
-		for _, r := range agent.Reasoners {
-			source := r.ProposedTags
-			if len(source) == 0 {
-				source = r.Tags
-			}
-			for _, t := range source {
-				if _, exists := proposedSeen[t]; !exists {
-					proposedSeen[t] = struct{}{}
-					agent.ProposedTags = append(agent.ProposedTags, t)
-				}
-			}
-		}
-		for _, sk := range agent.Skills {
-			source := sk.ProposedTags
-			if len(source) == 0 {
-				source = sk.Tags
-			}
-			for _, t := range source {
-				if _, exists := proposedSeen[t]; !exists {
-					proposedSeen[t] = struct{}{}
-					agent.ProposedTags = append(agent.ProposedTags, t)
-				}
-			}
-		}
-		types.HydrateAgentSessions(agent)
-		for _, session := range agent.Sessions {
-			source := session.ProposedTags
-			if len(source) == 0 {
-				source = session.Tags
-			}
-			for _, t := range source {
-				if _, exists := proposedSeen[t]; !exists {
-					proposedSeen[t] = struct{}{}
-					agent.ProposedTags = append(agent.ProposedTags, t)
-				}
-			}
-		}
-	}
-}
-
-// ============================================================================
-// Access Policy Storage
-// ============================================================================
-
-// GetAccessPolicies retrieves all enabled access policies, sorted by priority descending.
-func (ls *LocalStorage) GetAccessPolicies(ctx context.Context) ([]*types.AccessPolicy, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get access policies: %w", err)
-	}
-
-	query := `
-		SELECT id, name, caller_tags, target_tags, allow_functions, deny_functions,
-		       constraints, action, priority, enabled, description, created_at, updated_at
-		FROM access_policies WHERE enabled = true ORDER BY priority DESC, created_at DESC`
-
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to get access policies: %w", err)
-	}
-	defer rows.Close()
-
-	var policies []*types.AccessPolicy
-	for rows.Next() {
-		if err := ctx.Err(); err != nil {
-			return nil, fmt.Errorf("context cancelled during scan: %w", err)
-		}
-
-		policy, err := scanAccessPolicy(rows)
-		if err != nil {
-			return nil, err
-		}
-		policies = append(policies, policy)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, fmt.Errorf("error iterating access policies: %w", err)
-	}
-
-	return policies, nil
-}
-
-// GetAccessPolicyByID retrieves a single access policy by its ID.
-func (ls *LocalStorage) GetAccessPolicyByID(ctx context.Context, id int64) (*types.AccessPolicy, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get access policy: %w", err)
-	}
-
-	query := `
-		SELECT id, name, caller_tags, target_tags, allow_functions, deny_functions,
-		       constraints, action, priority, enabled, description, created_at, updated_at
-		FROM access_policies WHERE id = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, id)
-
-	policy := &types.AccessPolicy{}
-	var callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON string
-	var description sql.NullString
-
-	err := row.Scan(
-		&policy.ID, &policy.Name, &callerTagsJSON, &targetTagsJSON,
-		&allowFuncsJSON, &denyFuncsJSON, &constraintsJSON,
-		&policy.Action, &policy.Priority, &policy.Enabled, &description,
-		&policy.CreatedAt, &policy.UpdatedAt,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("access policy with ID %d not found: %w", id, err)
-	}
-
-	if description.Valid {
-		policy.Description = &description.String
-	}
-	if err := unmarshalAccessPolicyJSON(policy, callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal access policy %d: %w", id, err)
-	}
-
-	return policy, nil
-}
-
-// CreateAccessPolicy creates a new access policy.
-func (ls *LocalStorage) CreateAccessPolicy(ctx context.Context, policy *types.AccessPolicy) error {
-	if ls.mode == "postgres" {
-		return ls.createAccessPolicyPostgres(ctx, policy)
-	}
-
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during create access policy: %w", err)
-	}
-
-	callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON, err := marshalAccessPolicyJSON(policy)
-	if err != nil {
-		return fmt.Errorf("failed to marshal access policy fields: %w", err)
-	}
-
-	query := `
-		INSERT INTO access_policies (
-			name, caller_tags, target_tags, allow_functions, deny_functions,
-			constraints, action, priority, enabled, description, created_at, updated_at
-		) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)`
-
-	result, err := ls.db.ExecContext(ctx, query,
-		policy.Name, callerTagsJSON, targetTagsJSON,
-		allowFuncsJSON, denyFuncsJSON, constraintsJSON,
-		policy.Action, policy.Priority, policy.Enabled, policy.Description,
-		policy.CreatedAt, policy.UpdatedAt,
-	)
-	if err != nil {
-		if strings.Contains(err.Error(), "UNIQUE constraint") {
-			return fmt.Errorf("access policy with name %q already exists", policy.Name)
-		}
-		return fmt.Errorf("failed to create access policy: %w", err)
-	}
-
-	id, err := result.LastInsertId()
-	if err == nil {
-		policy.ID = id
-	}
-
-	return nil
-}
-
-// createAccessPolicyPostgres creates an access policy using PostgreSQL's RETURNING clause.
-func (ls *LocalStorage) createAccessPolicyPostgres(ctx context.Context, policy *types.AccessPolicy) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during create access policy: %w", err)
-	}
-
-	callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON, err := marshalAccessPolicyJSON(policy)
-	if err != nil {
-		return fmt.Errorf("failed to marshal access policy fields: %w", err)
-	}
-
-	query := `
-		INSERT INTO access_policies (
-			name, caller_tags, target_tags, allow_functions, deny_functions,
-			constraints, action, priority, enabled, description, created_at, updated_at
-		) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11, $12)
-		RETURNING id`
-
-	row := ls.db.DB.QueryRowContext(ctx, query,
-		policy.Name, callerTagsJSON, targetTagsJSON,
-		allowFuncsJSON, denyFuncsJSON, constraintsJSON,
-		policy.Action, policy.Priority, policy.Enabled, policy.Description,
-		policy.CreatedAt, policy.UpdatedAt,
-	)
-
-	if err := row.Scan(&policy.ID); err != nil {
-		if strings.Contains(err.Error(), "duplicate key") {
-			return fmt.Errorf("access policy with name %q already exists", policy.Name)
-		}
-		return fmt.Errorf("failed to create access policy: %w", err)
-	}
-
-	return nil
-}
-
-// UpdateAccessPolicy updates an existing access policy.
-func (ls *LocalStorage) UpdateAccessPolicy(ctx context.Context, policy *types.AccessPolicy) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during update access policy: %w", err)
-	}
-
-	callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON, err := marshalAccessPolicyJSON(policy)
-	if err != nil {
-		return fmt.Errorf("failed to marshal access policy fields: %w", err)
-	}
-
-	query := `
-		UPDATE access_policies SET
-			name = ?, caller_tags = ?, target_tags = ?, allow_functions = ?,
-			deny_functions = ?, constraints = ?, action = ?, priority = ?,
-			enabled = ?, description = ?, updated_at = ?
-		WHERE id = ?`
-
-	result, err := ls.db.ExecContext(ctx, query,
-		policy.Name, callerTagsJSON, targetTagsJSON,
-		allowFuncsJSON, denyFuncsJSON, constraintsJSON,
-		policy.Action, policy.Priority, policy.Enabled, policy.Description,
-		policy.UpdatedAt, policy.ID,
-	)
-	if err != nil {
-		return fmt.Errorf("failed to update access policy: %w", err)
-	}
-
-	rowsAffected, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to get rows affected: %w", err)
-	}
-	if rowsAffected == 0 {
-		return fmt.Errorf("access policy with ID %d not found", policy.ID)
-	}
-
-	return nil
-}
-
-// DeleteAccessPolicy deletes an access policy by ID.
-func (ls *LocalStorage) DeleteAccessPolicy(ctx context.Context, id int64) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during delete access policy: %w", err)
-	}
-
-	query := `DELETE FROM access_policies WHERE id = ?`
-
-	result, err := ls.db.ExecContext(ctx, query, id)
-	if err != nil {
-		return fmt.Errorf("failed to delete access policy: %w", err)
-	}
-
-	rows, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to get rows affected: %w", err)
-	}
-	if rows == 0 {
-		return fmt.Errorf("access policy with ID %d not found", id)
-	}
-
-	return nil
-}
-
-// scanAccessPolicy scans a row into an AccessPolicy struct.
-func scanAccessPolicy(rows *sql.Rows) (*types.AccessPolicy, error) {
-	policy := &types.AccessPolicy{}
-	var callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON string
-	var description sql.NullString
-
-	err := rows.Scan(
-		&policy.ID, &policy.Name, &callerTagsJSON, &targetTagsJSON,
-		&allowFuncsJSON, &denyFuncsJSON, &constraintsJSON,
-		&policy.Action, &policy.Priority, &policy.Enabled, &description,
-		&policy.CreatedAt, &policy.UpdatedAt,
-	)
-	if err != nil {
-		return nil, fmt.Errorf("failed to scan access policy: %w", err)
-	}
-
-	if description.Valid {
-		policy.Description = &description.String
-	}
-	if err := unmarshalAccessPolicyJSON(policy, callerTagsJSON, targetTagsJSON, allowFuncsJSON, denyFuncsJSON, constraintsJSON); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal access policy %d: %w", policy.ID, err)
-	}
-
-	return policy, nil
-}
-
-// unmarshalAccessPolicyJSON populates the JSON fields of an AccessPolicy.
-// Returns an error if any JSON field cannot be deserialized, preventing
-// corrupted data from silently producing empty policy rules.
-func unmarshalAccessPolicyJSON(policy *types.AccessPolicy, callerTags, targetTags, allowFuncs, denyFuncs, constraints string) error {
-	if callerTags != "" {
-		if err := json.Unmarshal([]byte(callerTags), &policy.CallerTags); err != nil {
-			return fmt.Errorf("failed to unmarshal caller_tags: %w", err)
-		}
-	}
-	if targetTags != "" {
-		if err := json.Unmarshal([]byte(targetTags), &policy.TargetTags); err != nil {
-			return fmt.Errorf("failed to unmarshal target_tags: %w", err)
-		}
-	}
-	if allowFuncs != "" {
-		if err := json.Unmarshal([]byte(allowFuncs), &policy.AllowFunctions); err != nil {
-			return fmt.Errorf("failed to unmarshal allow_functions: %w", err)
-		}
-	}
-	if denyFuncs != "" {
-		if err := json.Unmarshal([]byte(denyFuncs), &policy.DenyFunctions); err != nil {
-			return fmt.Errorf("failed to unmarshal deny_functions: %w", err)
-		}
-	}
-	if constraints != "" {
-		if err := json.Unmarshal([]byte(constraints), &policy.Constraints); err != nil {
-			return fmt.Errorf("failed to unmarshal constraints: %w", err)
-		}
-	}
-	return nil
-}
-
-// marshalAccessPolicyJSON serializes the JSON fields of an AccessPolicy for storage.
-func marshalAccessPolicyJSON(policy *types.AccessPolicy) (callerTags, targetTags, allowFuncs, denyFuncs, constraints string, err error) {
-	ct, err := json.Marshal(policy.CallerTags)
-	if err != nil {
-		return "", "", "", "", "", fmt.Errorf("caller_tags: %w", err)
-	}
-	tt, err := json.Marshal(policy.TargetTags)
-	if err != nil {
-		return "", "", "", "", "", fmt.Errorf("target_tags: %w", err)
-	}
-	af, err := json.Marshal(policy.AllowFunctions)
-	if err != nil {
-		return "", "", "", "", "", fmt.Errorf("allow_functions: %w", err)
-	}
-	df, err := json.Marshal(policy.DenyFunctions)
-	if err != nil {
-		return "", "", "", "", "", fmt.Errorf("deny_functions: %w", err)
-	}
-	cn, err := json.Marshal(policy.Constraints)
-	if err != nil {
-		return "", "", "", "", "", fmt.Errorf("constraints: %w", err)
-	}
-	return string(ct), string(tt), string(af), string(df), string(cn), nil
-}
-
-// ========== Agent Tag VC operations ==========
-
-// StoreAgentTagVC stores or replaces an agent's tag VC.
-func (ls *LocalStorage) StoreAgentTagVC(ctx context.Context, agentID, agentDID, vcID, vcDocument, signature string, issuedAt time.Time, expiresAt *time.Time) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during store agent tag VC: %w", err)
-	}
-
-	query := `
-		INSERT INTO agent_tag_vcs (agent_id, agent_did, vc_id, vc_document, signature, issued_at, expires_at, created_at, updated_at)
-		VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?)
-		ON CONFLICT(agent_id) DO UPDATE SET
-			agent_did = excluded.agent_did,
-			vc_id = excluded.vc_id,
-			vc_document = excluded.vc_document,
-			signature = excluded.signature,
-			issued_at = excluded.issued_at,
-			expires_at = excluded.expires_at,
-			revoked_at = NULL,
-			updated_at = excluded.updated_at`
-
-	now := time.Now()
-	_, err := ls.db.ExecContext(ctx, query, agentID, agentDID, vcID, vcDocument, signature, issuedAt, expiresAt, now, now)
-	if err != nil {
-		return fmt.Errorf("failed to store agent tag VC: %w", err)
-	}
-	return nil
-}
-
-// GetAgentTagVC retrieves an agent's tag VC record.
-func (ls *LocalStorage) GetAgentTagVC(ctx context.Context, agentID string) (*types.AgentTagVCRecord, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during get agent tag VC: %w", err)
-	}
-
-	query := `
-		SELECT id, agent_id, agent_did, vc_id, vc_document, signature, issued_at, expires_at, revoked_at
-		FROM agent_tag_vcs WHERE agent_id = ?`
-
-	row := ls.db.QueryRowContext(ctx, query, agentID)
-
-	record := &types.AgentTagVCRecord{}
-	var expiresAt, revokedAt sql.NullTime
-	var signature sql.NullString
-
-	err := row.Scan(
-		&record.ID, &record.AgentID, &record.AgentDID, &record.VCID,
-		&record.VCDocument, &signature, &record.IssuedAt, &expiresAt, &revokedAt,
-	)
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, fmt.Errorf("agent tag VC not found for agent %s", agentID)
-		}
-		return nil, fmt.Errorf("failed to get agent tag VC: %w", err)
-	}
-
-	if signature.Valid {
-		record.Signature = signature.String
-	}
-	if expiresAt.Valid {
-		record.ExpiresAt = &expiresAt.Time
-	}
-	if revokedAt.Valid {
-		record.RevokedAt = &revokedAt.Time
-	}
-
-	return record, nil
-}
-
-// ListAgentTagVCs returns all non-revoked agent tag VCs.
-func (ls *LocalStorage) ListAgentTagVCs(ctx context.Context) ([]*types.AgentTagVCRecord, error) {
-	if err := ctx.Err(); err != nil {
-		return nil, fmt.Errorf("context cancelled during list agent tag VCs: %w", err)
-	}
-
-	query := `
-		SELECT id, agent_id, agent_did, vc_id, vc_document, signature, issued_at, expires_at, revoked_at
-		FROM agent_tag_vcs WHERE revoked_at IS NULL`
-
-	rows, err := ls.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, fmt.Errorf("failed to list agent tag VCs: %w", err)
-	}
-	defer rows.Close()
-
-	var records []*types.AgentTagVCRecord
-	for rows.Next() {
-		record := &types.AgentTagVCRecord{}
-		var expiresAt, revokedAt sql.NullTime
-		var signature sql.NullString
-
-		if err := rows.Scan(
-			&record.ID, &record.AgentID, &record.AgentDID, &record.VCID,
-			&record.VCDocument, &signature, &record.IssuedAt, &expiresAt, &revokedAt,
-		); err != nil {
-			return nil, fmt.Errorf("failed to scan agent tag VC: %w", err)
-		}
-
-		if signature.Valid {
-			record.Signature = signature.String
-		}
-		if expiresAt.Valid {
-			record.ExpiresAt = &expiresAt.Time
-		}
-		if revokedAt.Valid {
-			record.RevokedAt = &revokedAt.Time
-		}
-		records = append(records, record)
-	}
-
-	return records, rows.Err()
-}
-
-// RevokeAgentTagVC marks an agent's tag VC as revoked.
-func (ls *LocalStorage) RevokeAgentTagVC(ctx context.Context, agentID string) error {
-	if err := ctx.Err(); err != nil {
-		return fmt.Errorf("context cancelled during revoke agent tag VC: %w", err)
-	}
-
-	query := `UPDATE agent_tag_vcs SET revoked_at = ?, updated_at = ? WHERE agent_id = ? AND revoked_at IS NULL`
-
-	now := time.Now()
-	result, err := ls.db.ExecContext(ctx, query, now, now, agentID)
-	if err != nil {
-		return fmt.Errorf("failed to revoke agent tag VC: %w", err)
-	}
-
-	rows, err := result.RowsAffected()
-	if err != nil {
-		return fmt.Errorf("failed to get rows affected: %w", err)
-	}
-	if rows == 0 {
-		return fmt.Errorf("no active agent tag VC found for agent %s", agentID)
-	}
-
-	return nil
-}
